==> 00-data-exploration-revision.Rmd <==
---
title: "Supplementary material - Exploratory analyses"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
library(tidyverse)
library(tidybayes)
library(ggpubr)
library(bootstrap)
library(kableExtra)

RT_SHAPE_crit <- readRDS("RT_SHAPE_crit.RDS")
RT_SHAPE_trial <- readRDS("RT_SHAPE_trial.RDS")

RT_SHAPE_respond <- readRDS("RT_SHAPE_respond.RDS")
RT_SHAPE_respond_crit <- readRDS("RT_SHAPE_respond_crit.RDS")
RT_SHAPE_respond_trial <- readRDS("RT_SHAPE_respond_trial.RDS")

theta <- function(x,xdata,na.rm=T) {mean(xdata[x],na.rm=na.rm)}
ci.low <- function(x,na.rm=T) {
  mean(x,na.rm=na.rm) - quantile(bootstrap(1:length(x),1000,theta,x,na.rm=na.rm)$thetastar,.025,na.rm=na.rm)}
ci.high <- function(x,na.rm=T) {
  quantile(bootstrap(1:length(x),1000,theta,x,na.rm=na.rm)$thetastar,.975,na.rm=na.rm) - mean(x,na.rm=na.rm)}

# Import data
d <- read_csv("results-25-11-sentence.csv", locale = locale(encoding = 'ISO-8859-1'))

# Check number of data points per participant in order to remove incomplete submissions
table(d$submission_id, d$trial_number)

# Remove incomplete submissions and practice trials
d <- d %>% 
  filter(submission_id != 7950, submission_id != 7954) %>% 
  filter(trial_name != "practice_task_three")

# Convert variable 'trial_type' to factor
d$trial_type <- as.factor(d$trial_type)

# Clean data set
# Remove unnecessary variables
d2 <- d %>%
  filter(is.na(trial_type)) %>%
  select(-trial_type, -property_1, -property_2, -shape_1, -shape_2, -response_1, -response_2, -response_3, -display, -experiment_id) %>% 
  gather(Region, RT, -submission_id, -quantifier, -picture_type, -condition, -sentence, -response, -listNumber,
         -trial_number, -picture, -trial_name)

# Convert variable 'Region' to factor
d2$Region = factor(x = d2$Region,
                   levels = c("QUANT", "der", "SHAPE", "auf",
                              "dem", "Bild", "sind", "CRIT_3", "der_1",
                              "Box"))

# Convert to factor variables to be included in the models as categorical predictors
d2$submission_id <- as.factor(d2$submission_id)
d2$quantifier <- as.factor(d2$quantifier)
d2$condition <- as.factor(d2$condition)

# Check for outliers
# Check for outlier trials (= RTs larger/ smaller than 2.5 SD of the mean)
# Check for outlier participants (= total number of outlier trials larger than a third of all trials)
d2 <- d2 %>% group_by(Region, quantifier, condition) %>%
  mutate(outlier_trial = RT < (mean(RT) - 2.5*sd(RT)) | RT > (mean(RT) + 2.5*sd(RT)) ) %>%
  ungroup() %>% 
  group_by(submission_id) %>% 
  mutate(outlier_participant = mean(outlier_trial, na.rm = TRUE) > 0.3) %>% 
  ungroup() 

# Show outliers
show(paste0("Excluded trials: ", sum(d2$outlier_trial, na.rm = TRUE) ))
show(paste0("Excluded participants: ", sum(d2$outlier_participant) ))

# Save data set with removed outliers
d2 <- d2 %>% filter(outlier_trial == FALSE) %>% 
  select(-outlier_trial, -outlier_participant)

# Create variable 'logCon' (= logical conditions)
# Variable consists of the relevant combinations of 'quantifier' and 'condition', as per the experimental conditions
d2Logic <- d2 %>%
  filter(trial_name != "practice_task_three") %>% 
  mutate(logCon = case_when(quantifier == "some" & condition == "unbiased" ~ "Einige (Unbiased)",
                            quantifier == "some" & condition == "biased" ~ "Einige (Biased)",
                            quantifier == "some" & condition == "false" ~ "Einige (False)",
                            quantifier == "some" & condition == "underspecified" ~ "Einige (Infelict)",
                            quantifier == "all" & condition == "unbiased" ~ "Alle (Unbiased)",
                            quantifier == "all" & condition == "biased" ~ "Alle (Biased)",
                            quantifier == "all" & condition == "false" ~ "Alle (False)",
                            TRUE ~ "other")) %>% 
  select(-quantifier, -condition) %>% 
  filter(picture_type != "A" | logCon != "Einige (False)")

# Convert variable 'logCon' to factor
d2Logic$logCon  <- factor(x = d2Logic$logCon, levels = c("Einige (Biased)", "Einige (Unbiased)", "Einige (False)", "Einige (Infelict)", "Alle (Biased)", "Alle (Unbiased)", "Alle (False)"))

trial_means <- d2Logic %>%
  filter(Region == "SHAPE") %>%
  group_by(logCon, trial_name) %>%
  summarise(Mean = mean(RT), CILow = ci.low(RT), CIHigh = ci.high(RT)) %>%
  mutate(YMin = Mean-CILow, YMax = Mean+CIHigh)

block_labels <- c("task_three_1" = "Trial 1-21",
                  "task_three_2" = "Trial 22-42",
                  "task_three_3" = "Trial 43-63",
                  "task_three_4" = "Trial 64-84")

# Extract strings with shape terms from variable 'sentence' and store in a new variable called 'critical'
d2Logic$critical <- str_extract(d2Logic$sentence, "Quadrate|Dreiecke|Dreicke|Kreise")
d2Logic$critical <- gsub("Dreicke", "Dreiecke", d2Logic$critical)

# Convert variables 'critical' and 'trial_name' to factor
d2Logic$critical <- as.factor(d2Logic$critical)

shape_means <- d2Logic %>%
  filter(Region == "SHAPE") %>%
  group_by(logCon, critical) %>%
  summarise(Mean = mean(RT), CILow = ci.low(RT), CIHigh = ci.high(RT)) %>%
  mutate(YMin = Mean-CILow, YMax = Mean+CIHigh)

```
 
***

### Shape term effects

#### Empirical means

```{r fig.align='center', fig.height=4, echo=FALSE, warning=FALSE, message=FALSE}

pd <- position_dodge(width = 0.9)

d2Logic %>%
  filter(Region == "SHAPE") %>%
  group_by(Region, logCon, critical) %>%
  summarise(Mean = mean(RT), CILow = ci.low(RT), CIHigh = ci.high(RT)) %>%
  mutate(YMin = Mean-CILow, YMax = Mean+CIHigh) %>% 
  ggplot(aes(Region, Mean, color = logCon, group = logCon, fill = logCon)) +
  geom_errorbar(aes(ymin = YMin, ymax = YMax), width = .1, 
                size = 1.1, alpha = .8, position = pd) +
  geom_point(aes(shape = logCon), size = 3.5, alpha = .9, position = pd, color = "black", stroke = .9) +
  facet_wrap( ~ critical) +
  theme_minimal() +
  # facet_wrap( ~ logCon, nrow = 1, labeller = as_labeller(block_labels)) +
  theme(axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        strip.text.x = element_text(face = "bold", color = "black", size = 14, 
                                   vjust = .5,
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.title.y = element_text(face = "bold", size = 14, angle = 0, vjust = .5,
                                    margin = margin(t = 0, r = 15, b = 0, l = 0)),
        legend.background = element_rect(size = .5, linetype = "solid"),  
        plot.title = element_text(hjust = 0.5), panel.grid.major.x = element_blank(),
        legend.position = "bottom") +
  labs(y = "RTs") +
  scale_color_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), times = 2), guide = guide_legend(title = "Condition", title.hjust = 0.5,  byrow = T)) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), times = 2), guide = guide_legend(title = "Condition", title.hjust = 0.5, byrow = T)) +
  scale_shape_manual(values = c(21, 21, 21, 21, 22, 22, 22), guide = guide_legend(title = "Condition", title.hjust = 0.5,  byrow = T))

# ggsave("RT-shape-means.png", height = 4, dpi = "retina")

```

Empirical means in numbers for relevant conditions and numerical effects for raw RTs:

```{r echo=FALSE, warning=FALSE, message=FALSE}

d2Logic %>%
  mutate(`Shape term` = critical) %>% 
  filter(Region == "SHAPE") %>% 
  group_by(logCon, `Shape term`) %>% 
  summarize(mean_RT = mean((RT))) %>% 
  ungroup() %>% 
  filter(str_detect(logCon, "iased")) %>% 
  pivot_wider(id_cols = `Shape term`, names_from = logCon, values_from = mean_RT) %>% 
  mutate(
    Einige = `Einige (Unbiased)` - `Einige (Biased)`,
    Alle = `Alle (Unbiased)` - `Alle (Biased)`
  ) %>% 
  kable(escape = F, align = "c") %>%
  kable_styling(latex_options = "striped", full_width = F) %>% 
  add_header_above(c(" " = 1, "Condition" = 4, "Effect" = 2))

```

Empirical means in numbers for relevant conditions and numerical effects for log(RTs):

```{r echo=FALSE, warning=FALSE, message=FALSE}

d2Logic %>%
  mutate(`Shape term` = critical) %>% 
  filter(Region == "SHAPE") %>% 
  group_by(logCon, `Shape term`) %>% 
  summarize(mean_RT = mean(log(RT))) %>% 
  ungroup() %>% 
  filter(str_detect(logCon, "iased")) %>% 
  pivot_wider(id_cols = `Shape term`, names_from = logCon, values_from = mean_RT) %>% 
  mutate(
    Einige = `Einige (Unbiased)` - `Einige (Biased)`,
    Alle = `Alle (Unbiased)` - `Alle (Biased)`
  ) %>% 
  kable(escape = F, align = "c") %>%
  kable_styling(latex_options = "striped", full_width = F) %>% 
  add_header_above(c(" " = 1, "Condition" = 4, "Effect" = 2))

```

#### Hypothesis tests

'CI': 90%-CI for one-sided and 95%-CI for two-sided hypotheses.

'*': For one-sided hypotheses, the posterior probability exceeds 95%; for two-sided hypotheses, the value tested against lies outside the 95%-CI.

Posterior probabilities of point hypotheses assume equal prior probabilities.

```{r echo=FALSE, warning=FALSE, message=FALSE}

hypotheses <- 
  c("logConEinigeUnbiased + critical1 + logConEinigeUnbiased:critical1 > critical1",
  "logConEinigeUnbiased + critical2 + logConEinigeUnbiased:critical2 > critical2",
  "(logConEinigeUnbiased + critical2 + logConEinigeUnbiased:critical2) - (logConEinigeUnbiased + critical1 + logConEinigeUnbiased:critical1) > critical2 - critical1",
  "logConAlleUnbiased + critical1 + logConAlleUnbiased:critical1 > logConAlleBiased + critical1 + logConAlleBiased:critical1",
  "logConAlleUnbiased + critical2 + logConAlleUnbiased:critical2 > logConAlleBiased + critical2 + logConAlleBiased:critical2",
  "(logConAlleUnbiased + critical2 + logConAlleUnbiased:critical2) - (logConAlleUnbiased + critical1 + logConAlleUnbiased:critical1) > (logConAlleBiased + critical2 + logConAlleBiased:critical2) - (logConAlleBiased + critical1 + logConAlleBiased:critical1)")

tests <- brms::hypothesis(RT_SHAPE_crit, hypotheses)

tests$hypothesis %>%
  mutate(Hypothesis = case_when(Hypothesis == "(logConEinigeUnbiased+critical1+logConEinigeUnbiased:critical1)-(critical1) > 0" ~ "[Dreiecke] Unbiased > Biased",
                        Hypothesis == "(logConEinigeUnbiased+critical2+logConEinigeUnbiased:critical2)-(critical2) > 0" ~ "[Kreise] Unbiased > Biased",
                      Hypothesis == "((logConEinigeUnbiased+critical2+logConEinigeUnbiased:critical2)-(logConEinigeUnbiased+critical1+logConEinigeUnbiased:critical1))-(critical2-critical1) > 0"  ~ "[Quadrate] Unbiased > Biased",
                      Hypothesis == "(logConAlleUnbiased+critical1+logConAlleUnbiased:critical1)-(logConAlleBiased+critical1+logConAlleBiased:critical1) > 0" ~ "[Dreiecke] Unbiased > Biased",
                      Hypothesis == "(logConAlleUnbiased+critical2+logConAlleUnbiased:critical2)-(logConAlleBiased+critical2+logConAlleBiased:critical2) > 0" ~ "[Kreise] Unbiased > Biased",
                      Hypothesis == "((logConAlleUnbiased+critical2+logConAlleUnbiased:critical2)-(logConAlleUnbiased+critical1+logConAlleUnbiased:critical1))-((logConAlleBiased+critical2+logConAlleBiased:critical2)-(logConAlleBiased+critical1+logConAlleBiased:critical1)) > 0" ~ "[Quadrate] Unbiased > Biased")) %>%
  # select(1:7) %>% 
  mutate_if(is.numeric, ~ round(., 2)) %>% 
  kable(escape = F, align = "c") %>%
  kable_styling(latex_options = "striped") %>% 
  pack_rows("Einige", 1, 3) %>%
  pack_rows("Alle", 4, 6)

```

#### Model (condition x critical word)

```{r fig.align='center', fig.height=6, fig.width=10, echo=FALSE, warning=FALSE, message=FALSE}

m_crit_einige <- RT_SHAPE_crit %>%
  spread_draws(b_Intercept,
               b_critical1,
               b_critical2,
               b_logConEinigeUnbiased,
               `b_logConEinigeUnbiased:critical1`,
               `b_logConEinigeUnbiased:critical2`,
               b_logConEinigeInfelict,
               `b_logConEinigeInfelict:critical1`, 
               `b_logConEinigeInfelict:critical2`,
               b_logConEinigeFalse,
               `b_logConEinigeFalse:critical1`, 
               `b_logConEinigeFalse:critical2`) %>%
  mutate(b_critical1_new = b_Intercept + b_critical2,
         b_critical2_new = b_Intercept + b_critical1,
         b_logConEinigeUnbiased_new = b_Intercept + b_logConEinigeUnbiased,
         `b_logConEinigeUnbiased:critical1_new` = b_Intercept + b_logConEinigeUnbiased + b_critical2 + `b_logConEinigeUnbiased:critical2`,
         `b_logConEinigeUnbiased:critical2_new` = b_Intercept + b_logConEinigeUnbiased + b_critical1 + `b_logConEinigeUnbiased:critical1`,
         b_logConEinigeInfelict_new = b_Intercept + b_logConEinigeInfelict,
         `b_logConEinigeInfelict:critical1_new` = b_Intercept + b_logConEinigeInfelict + b_critical2 + `b_logConEinigeInfelict:critical2`,
         `b_logConEinigeInfelict:critical2_new` = b_Intercept + b_logConEinigeInfelict + b_critical1 + `b_logConEinigeInfelict:critical1`,
         b_logConEinigeFalse_new = b_Intercept + b_logConEinigeFalse,
         `b_logConEinigeFalse:critical1_new` = b_Intercept + b_logConEinigeFalse + b_critical2 + `b_logConEinigeFalse:critical2`,
         `b_logConEinigeFalse:critical2_new` = b_Intercept + b_logConEinigeFalse + b_critical1 + `b_logConEinigeFalse:critical1`) %>%
  gather_draws(b_Intercept,
               b_critical1_new,
               b_critical2_new,
               b_logConEinigeUnbiased_new,
               `b_logConEinigeUnbiased:critical1_new`,
               `b_logConEinigeUnbiased:critical2_new`,
               b_logConEinigeInfelict_new,
               `b_logConEinigeInfelict:critical1_new`, 
               `b_logConEinigeInfelict:critical2_new`,
               b_logConEinigeFalse_new,
               `b_logConEinigeFalse:critical1_new`, 
               `b_logConEinigeFalse:critical2_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_Intercept",
               "b_critical1_new",
               "b_critical2_new",
               "b_logConEinigeUnbiased_new",
               "b_logConEinigeUnbiased:critical1_new",
               "b_logConEinigeUnbiased:critical2_new",
               "b_logConEinigeFalse_new",
               "b_logConEinigeFalse:critical1_new", 
               "b_logConEinigeFalse:critical2_new",
               "b_logConEinigeInfelict_new",
               "b_logConEinigeInfelict:critical1_new", 
               "b_logConEinigeInfelict:critical2_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  scale_y_discrete(labels = c('Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate')) +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(3,3,3,3)),
                    breaks = c("b_Intercept",
                               "b_logConEinigeUnbiased_new",
                               "b_logConEinigeFalse_new",
                               "b_logConEinigeInfelict_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False",
                               "Infelicitous")) +
  ggtitle("Einige")

m_crit_alle <- RT_SHAPE_crit %>%
  spread_draws(b_Intercept,
               b_critical2,
               b_critical1,
               b_logConAlleBiased,
               `b_logConAlleBiased:critical1`,
               `b_logConAlleBiased:critical2`,
               b_logConAlleUnbiased,
               `b_logConAlleUnbiased:critical1`,
               `b_logConAlleUnbiased:critical2`,
               b_logConAlleFalse,
               `b_logConAlleFalse:critical2`, 
               `b_logConAlleFalse:critical1`) %>%
  mutate(b_logConAlleBiased_new = b_Intercept + b_logConAlleBiased,
         `b_logConAlleBiased:critical1_new` = b_Intercept + b_logConAlleBiased + b_critical2 + `b_logConAlleBiased:critical2`,
         `b_logConAlleBiased:critical2_new` = b_Intercept + b_logConAlleBiased + b_critical1 + `b_logConAlleBiased:critical1`,
         b_logConAlleUnbiased_new = b_Intercept + b_logConAlleUnbiased,
         `b_logConAlleUnbiased:critical1_new` = b_Intercept + b_logConAlleUnbiased + b_critical2 + `b_logConAlleUnbiased:critical2`,
         `b_logConAlleUnbiased:critical2_new` = b_Intercept + b_logConAlleUnbiased + b_critical1 + `b_logConAlleUnbiased:critical1`,
         b_logConAlleFalse_new = b_Intercept + b_logConAlleFalse,
         `b_logConAlleFalse:critical1_new` = b_Intercept + b_logConAlleFalse + b_critical2 + `b_logConAlleFalse:critical2`,
         `b_logConAlleFalse:critical2_new` = b_Intercept + b_logConAlleFalse + b_critical1 + `b_logConAlleFalse:critical1`) %>% 
  gather_draws(b_logConAlleBiased_new,
               `b_logConAlleBiased:critical2_new`,
               `b_logConAlleBiased:critical1_new`,
               b_logConAlleUnbiased_new,
               `b_logConAlleUnbiased:critical2_new`,
               `b_logConAlleUnbiased:critical1_new`,
               b_logConAlleFalse_new,
               `b_logConAlleFalse:critical2_new`, 
               `b_logConAlleFalse:critical1_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_logConAlleBiased_new",
               "b_logConAlleBiased:critical1_new",
               "b_logConAlleBiased:critical2_new",
               "b_logConAlleUnbiased_new",
               "b_logConAlleUnbiased:critical1_new",
               "b_logConAlleUnbiased:critical2_new",
               "b_logConAlleFalse_new",
               "b_logConAlleFalse:critical1_new", 
               "b_logConAlleFalse:critical2_new"))) %>%
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  scale_y_discrete(labels = c('Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate'),
                   position = "right") +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        plot.title = element_text(hjust = 1,
                                  face = "bold",
                                  size = 14,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
   scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(3,3,3,3)),
                     breaks = c("b_logConAlleBiased_new",
                                "b_logConAlleUnbiased_new",
                                "b_logConAlleFalse_new"),
                     labels = c("Biased",
                                "Unbiased",
                                "False")) +
  ggtitle("Alle")

# `est-crit` <- 
  ggarrange(m_crit_einige, m_crit_alle, common.legend = T, legend = "bottom")

# ggsave("est-crit.png", height = 6, width = 10, dpi = "retina")

```

***

### Block effects

#### Empirical means

```{r fig.align='center', fig.height=5, width=12, echo=FALSE, warning=FALSE, message=FALSE}

pd <- position_dodge(width = 0.6)

d2Logic %>%
  filter(Region == "SHAPE") %>%
  group_by(logCon, trial_name) %>%
  summarise(Mean = mean(RT), CILow = ci.low(RT), CIHigh = ci.high(RT)) %>%
  mutate(YMin = Mean-CILow, YMax = Mean+CIHigh) %>% 
  ggplot(aes(trial_name, Mean, color = logCon, group = logCon, fill = logCon)) +
  geom_line(size = .6, linetype = "dashed", alpha = .5, position = pd) +
  geom_errorbar(aes(ymin = YMin, ymax = YMax), width = .1, 
                size = 1.1, alpha = .8, position = pd) +
  geom_point(aes(shape = logCon), size = 3.5, alpha = .9, position = pd, color = "black", stroke = .9) +
  theme_minimal() +
  # facet_wrap( ~ logCon, nrow = 1, labeller = as_labeller(block_labels)) +
  theme(axis.title.x = element_blank(), panel.spacing = unit(2, "lines"),
        axis.text.x = element_text(face = "bold", color = "black", size = 14, 
                                   vjust = .5,
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.title.y = element_text(face = "bold", size = 14, angle = 0, vjust = .5,
                                    margin = margin(t = 0, r = 15, b = 0, l = 0)),
        legend.background = element_rect(size = .5, linetype = "solid"),  
        plot.title = element_text(hjust = 0.5), panel.grid.major.x = element_blank(),
        legend.position = "bottom") +
  labs(y = "RTs") +
  scale_color_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), times = 2), guide = guide_legend(title = "Condition", title.hjust = 0.5,  byrow = T)) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), times = 2), guide = guide_legend(title = "Condition", title.hjust = 0.5, byrow = T)) +
  scale_shape_manual(values = c(21, 21, 21, 21, 22, 22, 22), 
                     guide = guide_legend(title = "Condition", 
                                          title.hjust = 0.5,  byrow = T)) +
  scale_x_discrete(position = "top", labels = block_labels)

# ggsave("RT-block-means.png", height = 5, width = 9, dpi = "retina")

```

Empirical means in numbers for relevant conditions and numerical effects for raw RTs:

```{r echo=FALSE, warning=FALSE, message=FALSE}

d2Logic %>%
  mutate(Trial = factor(trial_name, labels = c("1-21",
                                               "22-42",
                                               "43-63",
                                               "64-84"))) %>% 
  filter(Region == "SHAPE") %>% 
  group_by(logCon, Trial) %>% 
  summarize(mean_RT = mean((RT))) %>% 
  ungroup() %>% 
  filter(str_detect(logCon, "iased")) %>% 
  pivot_wider(id_cols = Trial, names_from = logCon, values_from = mean_RT) %>% 
  mutate(
    Einige = `Einige (Unbiased)` - `Einige (Biased)`,
    Alle = `Alle (Unbiased)` - `Alle (Biased)`
  ) %>% 
  kable(escape = F, align = "c") %>%
  kable_styling(latex_options = "striped", full_width = F) %>% 
  add_header_above(c(" " = 1, "Condition" = 4, "Effect" = 2))


```

Empirical means in numbers for relevant conditions and numerical effects for log(RTs):

```{r echo=FALSE, warning=FALSE, message=FALSE}

d2Logic %>%
  mutate(Trial = factor(trial_name, labels = c("1-21",
                                               "22-42",
                                               "43-63",
                                               "64-84"))) %>% 
  filter(Region == "SHAPE") %>% 
  group_by(logCon, Trial) %>% 
  summarize(mean_RT = mean(log(RT))) %>% 
  ungroup() %>% 
  filter(str_detect(logCon, "iased")) %>% 
  pivot_wider(id_cols = Trial, names_from = logCon, values_from = mean_RT) %>% 
  mutate(
    Einige = `Einige (Unbiased)` - `Einige (Biased)`,
    Alle = `Alle (Unbiased)` - `Alle (Biased)`
  )  %>% 
  kable(escape = F, align = "c") %>%
  kable_styling(latex_options = "striped", full_width = F) %>% 
  add_header_above(c(" " = 1, "Condition" = 4, "Effect" = 2))

```

#### Hypothesis tests

'CI': 90%-CI for one-sided and 95%-CI for two-sided hypotheses.

'*': For one-sided hypotheses, the posterior probability exceeds 95%; for two-sided hypotheses, the value tested against lies outside the 95%-CI.

Posterior probabilities of point hypotheses assume equal prior probabilities.

```{r fig.align='center', echo=FALSE, warning=FALSE, message=FALSE}

hypotheses <- 
  c("Intercept + logConEinigeUnbiased > Intercept",
    "Intercept + logConEinigeUnbiased + trial_nametask_three_2 + logConEinigeUnbiased:trial_nametask_three_2 > Intercept + trial_nametask_three_2",
    "Intercept + logConEinigeUnbiased + trial_nametask_three_3 + logConEinigeUnbiased:trial_nametask_three_3 > Intercept + trial_nametask_three_3",
    "Intercept + logConEinigeUnbiased + trial_nametask_three_4 + logConEinigeUnbiased:trial_nametask_three_4 > Intercept + trial_nametask_three_4",
    "Intercept + logConAlleUnbiased > Intercept + logConAlleBiased",
    "Intercept + logConAlleUnbiased + trial_nametask_three_2 + logConAlleUnbiased:trial_nametask_three_2 > Intercept + logConAlleBiased + trial_nametask_three_2 + logConAlleBiased:trial_nametask_three_2",
    "Intercept + logConAlleUnbiased + trial_nametask_three_3 + logConAlleUnbiased:trial_nametask_three_3 > Intercept + logConAlleBiased + trial_nametask_three_3 + logConAlleBiased:trial_nametask_three_3",
    "Intercept + logConAlleUnbiased + trial_nametask_three_4 + logConAlleUnbiased:trial_nametask_three_4 > Intercept + logConAlleBiased + trial_nametask_three_4 + logConAlleBiased:trial_nametask_three_4")

tests <- brms::hypothesis(RT_SHAPE_trial, hypotheses)

tests$hypothesis %>%
  mutate(Hypothesis = case_when(Hypothesis == "(Intercept+logConEinigeUnbiased)-(Intercept) > 0" ~ "[1-21] Unbiased > Biased",
                        Hypothesis == "(Intercept+logConEinigeUnbiased+trial_nametask_three_2+logConEinigeUnbiased:trial_nametask_three_2)-(Intercept+trial_nametask_three_2) > 0" ~ "[22-42] Unbiased > Biased",
                      Hypothesis == "(Intercept+logConEinigeUnbiased+trial_nametask_three_3+logConEinigeUnbiased:trial_nametask_three_3)-(Intercept+trial_nametask_three_3) > 0"  ~ "[43-63] Unbiased > Biased",
                      Hypothesis == "(Intercept+logConEinigeUnbiased+trial_nametask_three_4+logConEinigeUnbiased:trial_nametask_three_4)-(Intercept+trial_nametask_three_4) > 0" ~ "[64-84] Unbiased > Biased",
                      Hypothesis == "(Intercept+logConAlleUnbiased)-(Intercept+logConAlleBiased) > 0" ~ "[1-21] Unbiased > Biased",
                      Hypothesis == "(Intercept+logConAlleUnbiased+trial_nametask_three_2+logConAlleUnbiased:trial_nametask_three_2)-(Intercept+logConAlleBiased+trial_nametask_three_2+logConAlleBiased:trial_nametask_three_2) > 0" ~ "[22-42] Unbiased > Biased",
                   Hypothesis ==   "(Intercept+logConAlleUnbiased+trial_nametask_three_3+logConAlleUnbiased:trial_nametask_three_3)-(Intercept+logConAlleBiased+trial_nametask_three_3+logConAlleBiased:trial_nametask_three_3) > 0" ~ "[43-63] Unbiased > Biased",
                    Hypothesis ==  "(Intercept+logConAlleUnbiased+trial_nametask_three_4+logConAlleUnbiased:trial_nametask_three_4)-(Intercept+logConAlleBiased+trial_nametask_three_4+logConAlleBiased:trial_nametask_three_4) > 0" ~ "[64-84] Unbiased > Biased")) %>% 
  # select(1:7) %>% 
  mutate_if(is.numeric, ~ round(., 2)) %>% 
  kable(escape = F, align = "c") %>%
  kable_styling(latex_options = "striped") %>% 
  pack_rows("Einige", 1, 4) %>%
  pack_rows("Alle", 5, 8)

```

#### Model (condition x block)

```{r fig.align='center', fig.height=6, fig.width=10, echo=FALSE, warning=FALSE, message=FALSE}

m_trial_einige <- RT_SHAPE_trial %>%
  spread_draws(b_Intercept,
               b_trial_nametask_three_2,
               b_trial_nametask_three_3,
               b_trial_nametask_three_4,
               b_logConEinigeUnbiased,
               `b_logConEinigeUnbiased:trial_nametask_three_2`,
               `b_logConEinigeUnbiased:trial_nametask_three_3`,
               `b_logConEinigeUnbiased:trial_nametask_three_4`,
               b_logConEinigeInfelict,
               `b_logConEinigeInfelict:trial_nametask_three_2`, 
               `b_logConEinigeInfelict:trial_nametask_three_3`,
               `b_logConEinigeInfelict:trial_nametask_three_4`,
               b_logConEinigeFalse,
               `b_logConEinigeFalse:trial_nametask_three_2`, 
               `b_logConEinigeFalse:trial_nametask_three_3`,
               `b_logConEinigeFalse:trial_nametask_three_4`) %>%
  mutate(b_trial_nametask_three_2_new = b_Intercept + b_trial_nametask_three_2,
         b_trial_nametask_three_3_new = b_Intercept + b_trial_nametask_three_3,
         b_trial_nametask_three_4_new = b_Intercept + b_trial_nametask_three_4,
         b_logConEinigeUnbiased_new = b_Intercept + b_logConEinigeUnbiased,
         `b_logConEinigeUnbiased:trial_nametask_three_2_new` = b_Intercept + b_logConEinigeUnbiased + b_trial_nametask_three_2 + `b_logConEinigeUnbiased:trial_nametask_three_2`,
         `b_logConEinigeUnbiased:trial_nametask_three_3_new` = b_Intercept + b_logConEinigeUnbiased + b_trial_nametask_three_3 + `b_logConEinigeUnbiased:trial_nametask_three_3`,
         `b_logConEinigeUnbiased:trial_nametask_three_4_new` = b_Intercept + b_logConEinigeUnbiased + b_trial_nametask_three_4 + `b_logConEinigeUnbiased:trial_nametask_three_4`,
         b_logConEinigeInfelict_new = b_Intercept + b_logConEinigeInfelict,
         `b_logConEinigeInfelict:trial_nametask_three_2_new` = b_Intercept + b_logConEinigeInfelict + b_trial_nametask_three_2 + `b_logConEinigeInfelict:trial_nametask_three_2`,
         `b_logConEinigeInfelict:trial_nametask_three_3_new` = b_Intercept + b_logConEinigeInfelict + b_trial_nametask_three_3 + `b_logConEinigeInfelict:trial_nametask_three_3`,
         `b_logConEinigeInfelict:trial_nametask_three_4_new` = b_Intercept + b_logConEinigeInfelict + b_trial_nametask_three_4 + `b_logConEinigeInfelict:trial_nametask_three_4`,
         b_logConEinigeFalse_new = b_Intercept + b_logConEinigeFalse,
         `b_logConEinigeFalse:trial_nametask_three_2_new` = b_Intercept + b_logConEinigeFalse + b_trial_nametask_three_2 + `b_logConEinigeFalse:trial_nametask_three_2`,
         `b_logConEinigeFalse:trial_nametask_three_3_new` = b_Intercept + b_logConEinigeFalse + b_trial_nametask_three_3 + `b_logConEinigeFalse:trial_nametask_three_3`,
         `b_logConEinigeFalse:trial_nametask_three_4_new` = b_Intercept + b_logConEinigeFalse + b_trial_nametask_three_4 + `b_logConEinigeFalse:trial_nametask_three_4`) %>%
  gather_draws(b_Intercept,
               b_trial_nametask_three_2_new,
               b_trial_nametask_three_3_new,
               b_trial_nametask_three_4_new,
               b_logConEinigeUnbiased_new,
               `b_logConEinigeUnbiased:trial_nametask_three_2_new`,
               `b_logConEinigeUnbiased:trial_nametask_three_3_new`,
               `b_logConEinigeUnbiased:trial_nametask_three_4_new`,
               b_logConEinigeFalse_new,
               `b_logConEinigeFalse:trial_nametask_three_2_new`, 
               `b_logConEinigeFalse:trial_nametask_three_3_new`,
               `b_logConEinigeFalse:trial_nametask_three_4_new`,
                b_logConEinigeInfelict_new,
               `b_logConEinigeInfelict:trial_nametask_three_2_new`, 
               `b_logConEinigeInfelict:trial_nametask_three_3_new`,
               `b_logConEinigeInfelict:trial_nametask_three_4_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_Intercept",
               "b_trial_nametask_three_2_new",
               "b_trial_nametask_three_3_new",
               "b_trial_nametask_three_4_new",
               "b_logConEinigeUnbiased_new",
               "b_logConEinigeUnbiased:trial_nametask_three_2_new",
               "b_logConEinigeUnbiased:trial_nametask_three_3_new",
               "b_logConEinigeUnbiased:trial_nametask_three_4_new",
               "b_logConEinigeFalse_new",
               "b_logConEinigeFalse:trial_nametask_three_2_new", 
               "b_logConEinigeFalse:trial_nametask_three_3_new",
               "b_logConEinigeFalse:trial_nametask_three_4_new",
                "b_logConEinigeInfelict_new",
               "b_logConEinigeInfelict:trial_nametask_three_2_new", 
               "b_logConEinigeInfelict:trial_nametask_three_3_new",
               "b_logConEinigeInfelict:trial_nametask_three_4_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  scale_y_discrete(labels = c('Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84')) +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(4,4,4,4)),
                    breaks = c("b_Intercept",
                               "b_logConEinigeUnbiased_new",
                               "b_logConEinigeFalse_new",
                               "b_logConEinigeInfelict_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False",
                               "Infelicitous")) +
  ggtitle("Einige")

m_trial_alle <- RT_SHAPE_trial %>%
  spread_draws(b_Intercept,
               b_trial_nametask_three_2,
               b_trial_nametask_three_3,
               b_trial_nametask_three_4,
               b_logConAlleBiased,
               `b_logConAlleBiased:trial_nametask_three_2`, 
               `b_logConAlleBiased:trial_nametask_three_3`,
               `b_logConAlleBiased:trial_nametask_three_4`,
               b_logConAlleUnbiased,
               `b_logConAlleUnbiased:trial_nametask_three_2`,
               `b_logConAlleUnbiased:trial_nametask_three_3`,
               `b_logConAlleUnbiased:trial_nametask_three_4`,
               b_logConAlleFalse,
               `b_logConAlleFalse:trial_nametask_three_2`, 
               `b_logConAlleFalse:trial_nametask_three_3`,
               `b_logConAlleFalse:trial_nametask_three_4`) %>%
  mutate(b_logConAlleBiased_new = b_Intercept + b_logConAlleBiased,
         `b_logConAlleBiased:trial_nametask_three_2_new` = b_Intercept + b_logConAlleBiased + b_trial_nametask_three_2 + `b_logConAlleBiased:trial_nametask_three_2`,
         `b_logConAlleBiased:trial_nametask_three_3_new` = b_Intercept + b_logConAlleBiased + b_trial_nametask_three_3 + `b_logConAlleBiased:trial_nametask_three_3`,
         `b_logConAlleBiased:trial_nametask_three_4_new` = b_Intercept + b_logConAlleBiased + b_trial_nametask_three_4 + `b_logConAlleBiased:trial_nametask_three_4`,
         b_logConAlleUnbiased_new = b_Intercept + b_logConAlleUnbiased,
         `b_logConAlleUnbiased:trial_nametask_three_2_new` = b_Intercept + b_logConAlleUnbiased + b_trial_nametask_three_2 + `b_logConAlleUnbiased:trial_nametask_three_2`,
         `b_logConAlleUnbiased:trial_nametask_three_3_new` = b_Intercept + b_logConAlleUnbiased + b_trial_nametask_three_3 + `b_logConAlleUnbiased:trial_nametask_three_3`,
         `b_logConAlleUnbiased:trial_nametask_three_4_new` = b_Intercept + b_logConAlleUnbiased + b_trial_nametask_three_4 + `b_logConAlleUnbiased:trial_nametask_three_4`,
         b_logConAlleFalse_new = b_Intercept + b_logConAlleFalse,
         `b_logConAlleFalse:trial_nametask_three_2_new` = b_Intercept + b_logConAlleFalse + b_trial_nametask_three_2 + `b_logConAlleFalse:trial_nametask_three_2`,
         `b_logConAlleFalse:trial_nametask_three_3_new` = b_Intercept + b_logConAlleFalse + b_trial_nametask_three_3 + `b_logConAlleFalse:trial_nametask_three_3`,
         `b_logConAlleFalse:trial_nametask_three_4_new` = b_Intercept + b_logConAlleFalse + b_trial_nametask_three_4 + `b_logConAlleFalse:trial_nametask_three_4`) %>%
  gather_draws(b_logConAlleBiased_new,
               `b_logConAlleBiased:trial_nametask_three_2_new`, 
               `b_logConAlleBiased:trial_nametask_three_3_new`,
               `b_logConAlleBiased:trial_nametask_three_4_new`,
               b_logConAlleUnbiased_new,
               `b_logConAlleUnbiased:trial_nametask_three_2_new`,
               `b_logConAlleUnbiased:trial_nametask_three_3_new`,
               `b_logConAlleUnbiased:trial_nametask_three_4_new`,
               b_logConAlleFalse_new,
               `b_logConAlleFalse:trial_nametask_three_2_new`, 
               `b_logConAlleFalse:trial_nametask_three_3_new`,
               `b_logConAlleFalse:trial_nametask_three_4_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_logConAlleBiased_new",
               "b_logConAlleBiased:trial_nametask_three_2_new", 
               "b_logConAlleBiased:trial_nametask_three_3_new",
               "b_logConAlleBiased:trial_nametask_three_4_new",
               "b_logConAlleUnbiased_new",
               "b_logConAlleUnbiased:trial_nametask_three_2_new",
               "b_logConAlleUnbiased:trial_nametask_three_3_new",
               "b_logConAlleUnbiased:trial_nametask_three_4_new",
               "b_logConAlleFalse_new",
               "b_logConAlleFalse:trial_nametask_three_2_new", 
               "b_logConAlleFalse:trial_nametask_three_3_new",
               "b_logConAlleFalse:trial_nametask_three_4_new"))) %>%
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  scale_y_discrete(labels = c('Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84'),
                   position = "right") +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        plot.title = element_text(hjust = 1,
                                  face = "bold",
                                  size = 14,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 0, b = 0, l = 10))) +
   scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(4,4,4,4)),
                     breaks = c("b_logConAlleBiased_new",
                                "b_logConAlleUnbiased_new",
                                "b_logConAlleFalse_new"),
                     labels = c("Biased",
                                "Unbiased",
                                "False")) +
  ggtitle("Alle")

# `est-block` <- 
  ggarrange(m_trial_einige, m_trial_alle, common.legend = T, legend = "bottom")

# ggsave("est-block.png", height = 6, width = 10, dpi = "retina")

```

***

### Pragmatic profile effects

#### Hypothesis tests

'CI': 90%-CI for one-sided and 95%-CI for two-sided hypotheses.

'*': For one-sided hypotheses, the posterior probability exceeds 95%; for two-sided hypotheses, the value tested against lies outside the 95%-CI.

Posterior probabilities of point hypotheses assume equal prior probabilities.

```{r fig.align='center', echo=FALSE, warning=FALSE, message=FALSE}

hypotheses <- 
  c("Intercept + logConEinigeUnbiased > Intercept",
    "Intercept + logConEinigeUnbiased + respond_typePragmatic + logConEinigeUnbiased:respond_typePragmatic > Intercept + respond_typePragmatic",
    "Intercept + logConEinigeUnbiased + respond_typeMixed + logConEinigeUnbiased:respond_typeMixed > Intercept + respond_typeMixed",
    "Intercept + logConAlleUnbiased > Intercept + logConAlleBiased",
    "Intercept + logConAlleUnbiased + respond_typePragmatic + logConAlleUnbiased:respond_typePragmatic > Intercept + logConAlleBiased + respond_typePragmatic + logConAlleBiased:respond_typePragmatic",
    "Intercept + logConAlleUnbiased + respond_typeMixed + logConAlleUnbiased:respond_typeMixed > Intercept + logConAlleBiased + respond_typeMixed + logConAlleBiased:respond_typeMixed")

tests <- brms::hypothesis(RT_SHAPE_respond, hypotheses)

tests$hypothesis %>%
  mutate(Hypothesis = case_when(Hypothesis == "(Intercept+logConEinigeUnbiased)-(Intercept) > 0" ~ "[semantic] Unbiased > Biased",
                        Hypothesis == "(Intercept+logConEinigeUnbiased+respond_typePragmatic+logConEinigeUnbiased:respond_typePragmatic)-(Intercept+respond_typePragmatic) > 0" ~ "[pragmatic] Unbiased > Biased",
                      Hypothesis == "(Intercept+logConEinigeUnbiased+respond_typeMixed+logConEinigeUnbiased:respond_typeMixed)-(Intercept+respond_typeMixed) > 0"  ~ "[mixed] Unbiased > Biased",
                      Hypothesis == "(Intercept+logConAlleUnbiased)-(Intercept+logConAlleBiased) > 0" ~ "[semantic] Unbiased > Biased",
                      Hypothesis == "(Intercept+logConAlleUnbiased+respond_typePragmatic+logConAlleUnbiased:respond_typePragmatic)-(Intercept+logConAlleBiased+respond_typePragmatic+logConAlleBiased:respond_typePragmatic) > 0" ~ "[pragmatic] Unbiased > Biased",
                      Hypothesis == "(Intercept+logConAlleUnbiased+respond_typeMixed+logConAlleUnbiased:respond_typeMixed)-(Intercept+logConAlleBiased+respond_typeMixed+logConAlleBiased:respond_typeMixed) > 0" ~ "[mixed] Unbiased > Biased")) %>% 
 # select(1:7) %>%
  mutate_if(is.numeric, ~ round(., 2)) %>% 
  kable(escape = F, align = "c") %>%
  kable_styling(latex_options = "striped") %>% 
  pack_rows("Einige", 1, 3) %>%
  pack_rows("Alle", 4, 6)

```

#### Model 1 (condition x pragmatic profile)

```{r fig.align='center', fig.height=6, fig.width=10, echo=FALSE, warning=FALSE, message=FALSE}

m_respond_einige <- RT_SHAPE_respond %>%
  spread_draws(b_Intercept,
               b_logConEinigeUnbiased,
               b_logConEinigeInfelict,
               b_logConEinigeFalse,
               b_respond_typePragmatic,
               `b_logConEinigeUnbiased:respond_typePragmatic`,
               `b_logConEinigeInfelict:respond_typePragmatic`,
               `b_logConEinigeFalse:respond_typePragmatic`,
               b_respond_typeMixed,
               `b_logConEinigeUnbiased:respond_typeMixed`,
               `b_logConEinigeInfelict:respond_typeMixed`,
               `b_logConEinigeFalse:respond_typeMixed`) %>%
  mutate(b_logConEinigeUnbiased_new = b_Intercept + b_logConEinigeUnbiased,
         b_logConEinigeInfelict_new = b_Intercept + b_logConEinigeInfelict,
         b_logConEinigeFalse_new = b_Intercept + b_logConEinigeFalse,
         b_respond_typePragmatic_new = b_Intercept + b_respond_typePragmatic,
         `b_logConEinigeUnbiased:respond_typePragmatic_new` = b_Intercept + b_logConEinigeUnbiased + b_respond_typePragmatic + `b_logConEinigeUnbiased:respond_typePragmatic`,
         `b_logConEinigeInfelict:respond_typePragmatic_new` = b_Intercept + b_logConEinigeInfelict + b_respond_typePragmatic + `b_logConEinigeInfelict:respond_typePragmatic`,
         `b_logConEinigeFalse:respond_typePragmatic_new` = b_Intercept + b_logConEinigeFalse + b_respond_typePragmatic + `b_logConEinigeFalse:respond_typePragmatic`,
         b_respond_typeMixed_new = b_Intercept + b_respond_typeMixed,
         `b_logConEinigeUnbiased:respond_typeMixed_new` = b_Intercept + b_logConEinigeUnbiased + b_respond_typeMixed + `b_logConEinigeUnbiased:respond_typeMixed`,
         `b_logConEinigeInfelict:respond_typeMixed_new` = b_Intercept + b_logConEinigeInfelict + b_respond_typeMixed + `b_logConEinigeInfelict:respond_typeMixed`,
         `b_logConEinigeFalse:respond_typeMixed_new` = b_Intercept + b_logConEinigeFalse + b_respond_typeMixed + `b_logConEinigeFalse:respond_typeMixed`) %>%
  gather_draws(b_Intercept,
               b_logConEinigeUnbiased_new, 
           b_logConEinigeFalse_new,
           b_logConEinigeInfelict_new, 
           b_respond_typePragmatic_new,
           `b_logConEinigeUnbiased:respond_typePragmatic_new`,
           `b_logConEinigeFalse:respond_typePragmatic_new`,
           `b_logConEinigeInfelict:respond_typePragmatic_new`,
           b_respond_typeMixed_new,
           `b_logConEinigeUnbiased:respond_typeMixed_new`,
           `b_logConEinigeFalse:respond_typeMixed_new`,
            `b_logConEinigeInfelict:respond_typeMixed_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_Intercept",
                    "b_logConEinigeUnbiased_new", 
           "b_logConEinigeFalse_new",
           "b_logConEinigeInfelict_new", 
           "b_respond_typePragmatic_new",
           "b_logConEinigeUnbiased:respond_typePragmatic_new",
           "b_logConEinigeFalse:respond_typePragmatic_new",
           "b_logConEinigeInfelict:respond_typePragmatic_new",
           "b_respond_typeMixed_new",
           "b_logConEinigeUnbiased:respond_typeMixed_new",
           "b_logConEinigeFalse:respond_typeMixed_new",
           "b_logConEinigeInfelict:respond_typeMixed_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  scale_y_discrete(labels = c('[semantic] Biased',
                              'Unbiased', 
                              'False',
                              'Infelict', 
                              '[pragmatic] Biased',
                              'Unbiased',
                              'False',
                              'Infelict', 
                              '[mixed] Biased',
                              'Unbiased',
                              'False',
                              'Infelict')) +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                times = 3),
                    breaks = c("b_Intercept",
                               "b_logConEinigeUnbiased_new",
                               "b_logConEinigeFalse_new",
                               "b_logConEinigeInfelict_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False",
                               "Infelicitous")) +
  ggtitle("Einige")

m_respond_alle <- RT_SHAPE_respond %>%
  spread_draws(b_Intercept,
               b_logConAlleBiased,
               b_logConAlleUnbiased,
               b_logConAlleFalse,
               b_respond_typePragmatic,
               `b_logConAlleBiased:respond_typePragmatic`,
               `b_logConAlleUnbiased:respond_typePragmatic`,
               `b_logConAlleFalse:respond_typePragmatic`,
               b_respond_typeMixed,
               `b_logConAlleBiased:respond_typeMixed`,
               `b_logConAlleUnbiased:respond_typeMixed`,
               `b_logConAlleFalse:respond_typeMixed`) %>%
  mutate(b_logConAlleBiased_new = b_Intercept + b_logConAlleBiased,
         b_logConAlleUnbiased_new = b_Intercept + b_logConAlleUnbiased,
         b_logConAlleFalse_new = b_Intercept + b_logConAlleFalse,
         b_respond_typePragmatic_new = b_Intercept + b_respond_typePragmatic,
         `b_logConAlleBiased:respond_typePragmatic_new` = b_Intercept + b_logConAlleBiased + b_respond_typePragmatic + `b_logConAlleBiased:respond_typePragmatic`,
         `b_logConAlleUnbiased:respond_typePragmatic_new` = b_Intercept + b_logConAlleUnbiased + b_respond_typePragmatic + `b_logConAlleUnbiased:respond_typePragmatic`,
         `b_logConAlleFalse:respond_typePragmatic_new` = b_Intercept + b_logConAlleFalse + b_respond_typePragmatic + `b_logConAlleFalse:respond_typePragmatic`,
         b_respond_typeMixed_new = b_Intercept + b_respond_typeMixed,
         `b_logConAlleBiased:respond_typeMixed_new` = b_Intercept + b_logConAlleBiased + b_respond_typeMixed + `b_logConAlleBiased:respond_typeMixed`,
         `b_logConAlleUnbiased:respond_typeMixed_new` = b_Intercept + b_logConAlleUnbiased + b_respond_typeMixed + `b_logConAlleUnbiased:respond_typeMixed`,
         `b_logConAlleFalse:respond_typeMixed_new` = b_Intercept + b_logConAlleFalse + b_respond_typeMixed + `b_logConAlleFalse:respond_typeMixed`) %>%
  gather_draws(b_logConAlleBiased_new,
               b_logConAlleUnbiased_new, 
           b_logConAlleFalse_new,
           `b_logConAlleBiased:respond_typePragmatic_new`,
           `b_logConAlleUnbiased:respond_typePragmatic_new`,
           `b_logConAlleFalse:respond_typePragmatic_new`,
           `b_logConAlleBiased:respond_typeMixed_new`,
           `b_logConAlleUnbiased:respond_typeMixed_new`,
           `b_logConAlleFalse:respond_typeMixed_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_logConAlleBiased_new",
                    "b_logConAlleUnbiased_new",
                    "b_logConAlleFalse_new",
                    "b_logConAlleBiased:respond_typePragmatic_new",
                    "b_logConAlleUnbiased:respond_typePragmatic_new",
                    "b_logConAlleFalse:respond_typePragmatic_new",
                    "b_logConAlleBiased:respond_typeMixed_new",
                    "b_logConAlleUnbiased:respond_typeMixed_new",
                    "b_logConAlleFalse:respond_typeMixed_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  scale_y_discrete(labels = c('[semantic] Biased',
                              'Unbiased', 
                              'False',
                              '[pragmatic] Biased',
                              'Unbiased',
                              'False',
                              '[mixed] Biased',
                              'Unbiased',
                              'False'),
                   position = "right") +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        plot.title = element_text(hjust = 1,
                                  face = "bold",
                                  size = 14,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
   scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B"), times = 3),
                     breaks = c("b_logConAlleBiased_new",
                                "b_logConAlleUnbiased_new",
                                "b_logConAlleFalse_new"),
                     labels = c("Biased",
                                "Unbiased",
                                "False")) +
  ggtitle("Alle")

# `est-respond` <-
  ggarrange(m_respond_einige, m_respond_alle, common.legend = T, legend = "bottom")

# ggsave("est-respond.png", height = 6, width = 10, dpi = "retina")

```

#### Model 2 (condition x pragmatic profile x critical word)

```{r fig.align='center', fig.height=10, fig.width=10, echo=FALSE, warning=FALSE, message=FALSE}

m_crit_einige_sem <- RT_SHAPE_respond_crit %>%
  spread_draws(b_Intercept,
               b_critical1,
               b_critical2,
               b_logConEinigeUnbiased,
               `b_logConEinigeUnbiased:critical1`,
               `b_logConEinigeUnbiased:critical2`,
               b_logConEinigeInfelict,
               `b_logConEinigeInfelict:critical1`, 
               `b_logConEinigeInfelict:critical2`,
               b_logConEinigeFalse,
               `b_logConEinigeFalse:critical1`, 
               `b_logConEinigeFalse:critical2`) %>%
  mutate(b_critical1_new = b_Intercept + b_critical2,
         b_critical2_new = b_Intercept + b_critical1,
         b_logConEinigeUnbiased_new = b_Intercept + b_logConEinigeUnbiased,
         `b_logConEinigeUnbiased:critical1_new` = b_Intercept + b_logConEinigeUnbiased + b_critical2 + `b_logConEinigeUnbiased:critical2`,
         `b_logConEinigeUnbiased:critical2_new` = b_Intercept + b_logConEinigeUnbiased + b_critical1 + `b_logConEinigeUnbiased:critical1`,
         b_logConEinigeInfelict_new = b_Intercept + b_logConEinigeInfelict,
         `b_logConEinigeInfelict:critical1_new` = b_Intercept + b_logConEinigeInfelict + b_critical2 + `b_logConEinigeInfelict:critical2`,
         `b_logConEinigeInfelict:critical2_new` = b_Intercept + b_logConEinigeInfelict + b_critical1 + `b_logConEinigeInfelict:critical1`,
         b_logConEinigeFalse_new = b_Intercept + b_logConEinigeFalse,
         `b_logConEinigeFalse:critical1_new` = b_Intercept + b_logConEinigeFalse + b_critical2 + `b_logConEinigeFalse:critical2`,
         `b_logConEinigeFalse:critical2_new` = b_Intercept + b_logConEinigeFalse + b_critical1 + `b_logConEinigeFalse:critical1`) %>%
  gather_draws(b_Intercept,
               b_critical1_new,
               b_critical2_new,
               b_logConEinigeUnbiased_new,
               `b_logConEinigeUnbiased:critical1_new`,
               `b_logConEinigeUnbiased:critical2_new`,
               b_logConEinigeInfelict_new,
               `b_logConEinigeInfelict:critical1_new`, 
               `b_logConEinigeInfelict:critical2_new`,
               b_logConEinigeFalse_new,
               `b_logConEinigeFalse:critical1_new`, 
               `b_logConEinigeFalse:critical2_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_Intercept",
               "b_critical1_new",
               "b_critical2_new",
               "b_logConEinigeUnbiased_new",
               "b_logConEinigeUnbiased:critical1_new",
               "b_logConEinigeUnbiased:critical2_new",
               "b_logConEinigeFalse_new",
               "b_logConEinigeFalse:critical1_new", 
               "b_logConEinigeFalse:critical2_new",
               "b_logConEinigeInfelict_new",
               "b_logConEinigeInfelict:critical1_new", 
               "b_logConEinigeInfelict:critical2_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  scale_y_discrete(labels = c('Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate')) +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(3,3,3,3)),
                    breaks = c("b_Intercept",
                               "b_logConEinigeUnbiased_new",
                               "b_logConEinigeFalse_new",
                               "b_logConEinigeInfelict_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False",
                               "Infelicitous")) +
  ggtitle("Einige\nSemantic")

m_crit_einige_mix <- RT_SHAPE_respond_crit %>%
  spread_draws(b_Intercept,
               b_critical1,
               b_critical2,
               b_respond_typeMixed,
               `b_respond_typeMixed:critical1`,
               `b_respond_typeMixed:critical2`,
               b_logConEinigeUnbiased,
               `b_logConEinigeUnbiased:critical1`,
               `b_logConEinigeUnbiased:critical2`,
               `b_logConEinigeUnbiased:respond_typeMixed`,
               `b_logConEinigeUnbiased:respond_typeMixed:critical1`,
               `b_logConEinigeUnbiased:respond_typeMixed:critical2`,
               b_logConEinigeInfelict,
               `b_logConEinigeInfelict:critical1`, 
               `b_logConEinigeInfelict:critical2`,
               `b_logConEinigeInfelict:respond_typeMixed`,
               `b_logConEinigeInfelict:respond_typeMixed:critical1`, 
               `b_logConEinigeInfelict:respond_typeMixed:critical2`,
               b_logConEinigeFalse,
               `b_logConEinigeFalse:critical1`, 
               `b_logConEinigeFalse:critical2`,
               `b_logConEinigeFalse:respond_typeMixed`,
               `b_logConEinigeFalse:respond_typeMixed:critical1`, 
               `b_logConEinigeFalse:respond_typeMixed:critical2`) %>%
  mutate(b_respond_typeMixed_new = b_Intercept + b_respond_typeMixed,
         `b_respond_typeMixed:critical1_new` = b_Intercept + b_respond_typeMixed + b_critical2 + `b_respond_typeMixed:critical2`,
         `b_respond_typeMixed:critical2_new` = b_Intercept + b_respond_typeMixed + b_critical1 + `b_respond_typeMixed:critical1`,
         `b_logConEinigeUnbiased:respond_typeMixed_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeUnbiased + `b_logConEinigeUnbiased:respond_typeMixed`,
         `b_logConEinigeUnbiased:respond_typeMixed:critical1_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeUnbiased + b_critical2 + `b_logConEinigeUnbiased:critical2` + `b_logConEinigeUnbiased:respond_typeMixed` + `b_respond_typeMixed:critical2` + `b_logConEinigeUnbiased:respond_typeMixed:critical2`,
         `b_logConEinigeUnbiased:respond_typeMixed:critical2_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeUnbiased + b_critical1 + `b_logConEinigeUnbiased:critical1` + `b_logConEinigeUnbiased:respond_typeMixed` + `b_respond_typeMixed:critical1` + `b_logConEinigeUnbiased:respond_typeMixed:critical1`,
         `b_logConEinigeInfelict:respond_typeMixed_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeInfelict + `b_logConEinigeInfelict:respond_typeMixed`,
         `b_logConEinigeInfelict:respond_typeMixed:critical1_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeInfelict + b_critical2 + `b_logConEinigeInfelict:critical2` + `b_logConEinigeInfelict:respond_typeMixed` + `b_respond_typeMixed:critical2` + `b_logConEinigeInfelict:respond_typeMixed:critical2`,
         `b_logConEinigeInfelict:respond_typeMixed:critical2_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeInfelict + b_critical1 + `b_logConEinigeInfelict:critical1` + `b_logConEinigeInfelict:respond_typeMixed` + `b_respond_typeMixed:critical1` + `b_logConEinigeInfelict:respond_typeMixed:critical1`,
         `b_logConEinigeFalse:respond_typeMixed_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeFalse + `b_logConEinigeFalse:respond_typeMixed`,
         `b_logConEinigeFalse:respond_typeMixed:critical1_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeFalse + b_critical2 + `b_logConEinigeFalse:critical2` + `b_logConEinigeFalse:respond_typeMixed` + `b_respond_typeMixed:critical2` + `b_logConEinigeFalse:respond_typeMixed:critical2`,
         `b_logConEinigeFalse:respond_typeMixed:critical2_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeFalse + b_critical1 + `b_logConEinigeFalse:critical1` + `b_logConEinigeFalse:respond_typeMixed` + `b_respond_typeMixed:critical1` + `b_logConEinigeFalse:respond_typeMixed:critical1`) %>%
  gather_draws(b_respond_typeMixed_new,
               `b_respond_typeMixed:critical1_new`,
               `b_respond_typeMixed:critical2_new`,
               `b_logConEinigeUnbiased:respond_typeMixed_new`,
               `b_logConEinigeUnbiased:respond_typeMixed:critical1_new`,
               `b_logConEinigeUnbiased:respond_typeMixed:critical2_new`,
               `b_logConEinigeInfelict:respond_typeMixed_new`,
               `b_logConEinigeInfelict:respond_typeMixed:critical1_new`, 
               `b_logConEinigeInfelict:respond_typeMixed:critical2_new`,
               `b_logConEinigeFalse:respond_typeMixed_new`,
               `b_logConEinigeFalse:respond_typeMixed:critical1_new`, 
               `b_logConEinigeFalse:respond_typeMixed:critical2_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_respond_typeMixed_new",
               "b_respond_typeMixed:critical1_new",
               "b_respond_typeMixed:critical2_new",
               "b_logConEinigeUnbiased:respond_typeMixed_new",
               "b_logConEinigeUnbiased:respond_typeMixed:critical1_new",
               "b_logConEinigeUnbiased:respond_typeMixed:critical2_new",
               "b_logConEinigeFalse:respond_typeMixed_new",
               "b_logConEinigeFalse:respond_typeMixed:critical1_new", 
               "b_logConEinigeFalse:respond_typeMixed:critical2_new",
               "b_logConEinigeInfelict:respond_typeMixed_new",
               "b_logConEinigeInfelict:respond_typeMixed:critical1_new", 
               "b_logConEinigeInfelict:respond_typeMixed:critical2_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  hjust = .5,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_blank()) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(3,3,3,3)),
                    breaks = c("b_respond_typeMixed_new",
                               "b_logConEinigeUnbiased:respond_typeMixed_new",
                               "b_logConEinigeFalse:respond_typeMixed_new",
                               "b_logConEinigeInfelict:respond_typeMixed_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False",
                               "Infelicitous")) +
  ggtitle("\nMixed")

m_crit_einige_prag <- RT_SHAPE_respond_crit %>%
  spread_draws(b_Intercept,
               b_critical1,
               b_critical2,
               b_respond_typePragmatic,
               `b_respond_typePragmatic:critical1`,
               `b_respond_typePragmatic:critical2`,
               b_logConEinigeUnbiased,
               `b_logConEinigeUnbiased:critical1`,
               `b_logConEinigeUnbiased:critical2`,
               `b_logConEinigeUnbiased:respond_typePragmatic`,
               `b_logConEinigeUnbiased:respond_typePragmatic:critical1`,
               `b_logConEinigeUnbiased:respond_typePragmatic:critical2`,
               b_logConEinigeInfelict,
               `b_logConEinigeInfelict:critical1`, 
               `b_logConEinigeInfelict:critical2`,
               `b_logConEinigeInfelict:respond_typePragmatic`,
               `b_logConEinigeInfelict:respond_typePragmatic:critical1`, 
               `b_logConEinigeInfelict:respond_typePragmatic:critical2`,
               b_logConEinigeFalse,
               `b_logConEinigeFalse:critical1`, 
               `b_logConEinigeFalse:critical2`,
               `b_logConEinigeFalse:respond_typePragmatic`,
               `b_logConEinigeFalse:respond_typePragmatic:critical1`, 
               `b_logConEinigeFalse:respond_typePragmatic:critical2`) %>%
  mutate(b_respond_typePragmatic_new = b_Intercept + b_respond_typePragmatic,
         `b_respond_typePragmatic:critical1_new` = b_Intercept + b_respond_typePragmatic + b_critical2 + `b_respond_typePragmatic:critical2`,
         `b_respond_typePragmatic:critical2_new` = b_Intercept + b_respond_typePragmatic + b_critical1 + `b_respond_typePragmatic:critical1`,
         `b_logConEinigeUnbiased:respond_typePragmatic_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeUnbiased + `b_logConEinigeUnbiased:respond_typePragmatic`,
         `b_logConEinigeUnbiased:respond_typePragmatic:critical1_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeUnbiased + b_critical2 + `b_logConEinigeUnbiased:critical2` + `b_logConEinigeUnbiased:respond_typePragmatic` + `b_respond_typePragmatic:critical2` + `b_logConEinigeUnbiased:respond_typePragmatic:critical1`,
         `b_logConEinigeUnbiased:respond_typePragmatic:critical2_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeUnbiased + b_critical1 + `b_logConEinigeUnbiased:critical1` + `b_logConEinigeUnbiased:respond_typePragmatic` + `b_respond_typePragmatic:critical1` + `b_logConEinigeUnbiased:respond_typePragmatic:critical1`,
         `b_logConEinigeInfelict:respond_typePragmatic_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeInfelict + `b_logConEinigeInfelict:respond_typePragmatic`,
         `b_logConEinigeInfelict:respond_typePragmatic:critical1_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeInfelict + b_critical2 + `b_logConEinigeInfelict:critical2` + `b_logConEinigeInfelict:respond_typePragmatic` + `b_respond_typePragmatic:critical2` + `b_logConEinigeInfelict:respond_typePragmatic:critical2`,
         `b_logConEinigeInfelict:respond_typePragmatic:critical2_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeInfelict + b_critical1 + `b_logConEinigeInfelict:critical1` + `b_logConEinigeInfelict:respond_typePragmatic` + `b_respond_typePragmatic:critical1` + `b_logConEinigeInfelict:respond_typePragmatic:critical1`,
         `b_logConEinigeFalse:respond_typePragmatic_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeFalse + `b_logConEinigeFalse:respond_typePragmatic`,
         `b_logConEinigeFalse:respond_typePragmatic:critical1_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeFalse + b_critical2 + `b_logConEinigeFalse:critical2` + `b_logConEinigeFalse:respond_typePragmatic` + `b_respond_typePragmatic:critical2` + `b_logConEinigeFalse:respond_typePragmatic:critical2`,
         `b_logConEinigeFalse:respond_typePragmatic:critical2_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeFalse + b_critical1 + `b_logConEinigeFalse:critical1` + `b_logConEinigeFalse:respond_typePragmatic` + `b_respond_typePragmatic:critical1` + `b_logConEinigeFalse:respond_typePragmatic:critical1`) %>%
  gather_draws(b_respond_typePragmatic_new,
               `b_respond_typePragmatic:critical1_new`,
               `b_respond_typePragmatic:critical2_new`,
               `b_logConEinigeUnbiased:respond_typePragmatic_new`,
               `b_logConEinigeUnbiased:respond_typePragmatic:critical1_new`,
               `b_logConEinigeUnbiased:respond_typePragmatic:critical2_new`,
               `b_logConEinigeInfelict:respond_typePragmatic_new`,
               `b_logConEinigeInfelict:respond_typePragmatic:critical1_new`, 
               `b_logConEinigeInfelict:respond_typePragmatic:critical2_new`,
               `b_logConEinigeFalse:respond_typePragmatic_new`,
               `b_logConEinigeFalse:respond_typePragmatic:critical1_new`, 
               `b_logConEinigeFalse:respond_typePragmatic:critical2_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_respond_typePragmatic_new",
               "b_respond_typePragmatic:critical1_new",
               "b_respond_typePragmatic:critical2_new",
               "b_logConEinigeUnbiased:respond_typePragmatic_new",
               "b_logConEinigeUnbiased:respond_typePragmatic:critical1_new",
               "b_logConEinigeUnbiased:respond_typePragmatic:critical2_new",
               "b_logConEinigeFalse:respond_typePragmatic_new",
               "b_logConEinigeFalse:respond_typePragmatic:critical1_new", 
               "b_logConEinigeFalse:respond_typePragmatic:critical2_new",
               "b_logConEinigeInfelict:respond_typePragmatic_new",
               "b_logConEinigeInfelict:respond_typePragmatic:critical1_new", 
               "b_logConEinigeInfelict:respond_typePragmatic:critical2_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  scale_y_discrete(labels = c('Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate'),
                   position = "right") +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  hjust = 1,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(3,3,3,3)),
                    breaks = c("b_respond_typePragmatic_new",
                               "b_logConEinigeUnbiased:respond_typePragmatic_new",
                               "b_logConEinigeFalse:respond_typePragmatic_new",
                               "b_logConEinigeInfelict:respond_typePragmatic_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False",
                               "Infelicitous")) +
  ggtitle("\nPragmatic")

`est-crit-einige` <-
  ggarrange(m_crit_einige_sem, m_crit_einige_mix, m_crit_einige_prag, common.legend = T, legend = "bottom", ncol = 3)

# ggsave("est-crit-einige.png", height = 6, width = 10, dpi = "retina")

```


```{r fig.align='center', fig.height=10, fig.width=10, echo=FALSE, warning=FALSE, message=FALSE}

m_crit_alle_sem <- RT_SHAPE_respond_crit %>%
  spread_draws(b_Intercept,
               b_critical1,
               b_critical2,
               b_logConAlleBiased,
               `b_logConAlleBiased:critical1`,
               `b_logConAlleBiased:critical2`,
               b_logConAlleUnbiased,
               `b_logConAlleUnbiased:critical1`,
               `b_logConAlleUnbiased:critical2`,
               b_logConAlleFalse,
               `b_logConAlleFalse:critical1`, 
               `b_logConAlleFalse:critical2`) %>%
  mutate(b_logConAlleBiased_new = b_Intercept + b_logConAlleBiased,
         `b_logConAlleBiased:critical1_new` = b_Intercept + b_logConAlleBiased + b_critical2 + `b_logConAlleBiased:critical2`,
         `b_logConAlleBiased:critical2_new` = b_Intercept + b_logConAlleBiased + b_critical1 + `b_logConAlleBiased:critical1`,
         b_logConAlleUnbiased_new = b_Intercept + b_logConAlleUnbiased,
         `b_logConAlleUnbiased:critical1_new` = b_Intercept + b_logConAlleUnbiased + b_critical2 + `b_logConAlleUnbiased:critical2`,
         `b_logConAlleUnbiased:critical2_new` = b_Intercept + b_logConAlleUnbiased + b_critical1 + `b_logConAlleUnbiased:critical1`,
         b_logConAlleFalse_new = b_Intercept + b_logConAlleFalse,
         `b_logConAlleFalse:critical1_new` = b_Intercept + b_logConAlleFalse + b_critical2 + `b_logConAlleFalse:critical2`,
         `b_logConAlleFalse:critical2_new` = b_Intercept + b_logConAlleFalse + b_critical1 + `b_logConAlleFalse:critical1`) %>%
  gather_draws(b_logConAlleBiased_new,
               `b_logConAlleBiased:critical1_new`,
               `b_logConAlleBiased:critical2_new`,
               b_logConAlleUnbiased_new,
               `b_logConAlleUnbiased:critical1_new`,
               `b_logConAlleUnbiased:critical2_new`,
               b_logConAlleFalse_new,
               `b_logConAlleFalse:critical1_new`, 
               `b_logConAlleFalse:critical2_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_logConAlleBiased_new",
               "b_logConAlleBiased:critical1_new",
               "b_logConAlleBiased:critical2_new",
               "b_logConAlleUnbiased_new",
               "b_logConAlleUnbiased:critical1_new",
               "b_logConAlleUnbiased:critical2_new",
               "b_logConAlleFalse_new",
               "b_logConAlleFalse:critical1_new", 
               "b_logConAlleFalse:critical2_new",
               "b_logConAlleInfelict_new",
               "b_logConAlleInfelict:critical1_new", 
               "b_logConAlleInfelict:critical2_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  scale_y_discrete(labels = c('Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate')) +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(3,3,3,3)),
                    breaks = c("b_logConAlleBiased_new",
                               "b_logConAlleUnbiased_new",
                               "b_logConAlleFalse_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False")) +
  ggtitle("Alle\nSemantic")

m_crit_alle_mix <- RT_SHAPE_respond_crit %>%
  spread_draws(b_Intercept,
               b_critical1,
               b_critical2,
               b_respond_typeMixed,
               `b_respond_typeMixed:critical1`,
               `b_respond_typeMixed:critical2`,
               b_logConAlleBiased,
               `b_logConAlleBiased:critical1`, 
               `b_logConAlleBiased:critical2`,
               `b_logConAlleBiased:respond_typeMixed`,
               `b_logConAlleBiased:respond_typeMixed:critical1`, 
               `b_logConAlleBiased:respond_typeMixed:critical2`,
               b_logConAlleUnbiased,
               `b_logConAlleUnbiased:critical1`,
               `b_logConAlleUnbiased:critical2`,
               `b_logConAlleUnbiased:respond_typeMixed`,
               `b_logConAlleUnbiased:respond_typeMixed:critical1`,
               `b_logConAlleUnbiased:respond_typeMixed:critical2`,
               b_logConAlleFalse,
               `b_logConAlleFalse:critical1`, 
               `b_logConAlleFalse:critical2`,
               `b_logConAlleFalse:respond_typeMixed`,
               `b_logConAlleFalse:respond_typeMixed:critical1`, 
               `b_logConAlleFalse:respond_typeMixed:critical2`) %>%
  mutate(`b_logConAlleBiased:respond_typeMixed_new` = b_Intercept + b_logConAlleBiased + b_respond_typeMixed + `b_logConAlleBiased:respond_typeMixed`,
         `b_logConAlleBiased:respond_typeMixed:critical1_new` = b_Intercept + b_respond_typeMixed + b_logConAlleBiased + b_critical2 + `b_logConAlleBiased:respond_typeMixed` + `b_logConAlleBiased:critical2` + `b_respond_typeMixed:critical2` + `b_logConAlleBiased:respond_typeMixed:critical2`,
         `b_logConAlleBiased:respond_typeMixed:critical2_new` = b_Intercept + b_respond_typeMixed + b_logConAlleBiased + b_critical1 + `b_logConAlleBiased:respond_typeMixed` + `b_logConAlleBiased:critical1` + `b_respond_typeMixed:critical1` + `b_logConAlleBiased:respond_typeMixed:critical1`,
         `b_logConAlleUnbiased:respond_typeMixed_new` = b_Intercept + b_respond_typeMixed + b_logConAlleUnbiased + `b_logConAlleUnbiased:respond_typeMixed`,
         `b_logConAlleUnbiased:respond_typeMixed:critical1_new` = b_Intercept + b_respond_typeMixed + b_logConAlleUnbiased + b_critical2 + `b_logConAlleUnbiased:critical2` + `b_logConAlleUnbiased:respond_typeMixed` + `b_respond_typeMixed:critical2` + `b_logConAlleUnbiased:respond_typeMixed:critical2`,
         `b_logConAlleUnbiased:respond_typeMixed:critical2_new` = b_Intercept + b_respond_typeMixed + b_logConAlleUnbiased + b_critical1 + `b_logConAlleUnbiased:critical1` + `b_logConAlleUnbiased:respond_typeMixed` + `b_respond_typeMixed:critical1` + `b_logConAlleUnbiased:respond_typeMixed:critical1`,
         `b_logConAlleFalse:respond_typeMixed_new` = b_Intercept + b_respond_typeMixed + b_logConAlleFalse + `b_logConAlleFalse:respond_typeMixed`,
         `b_logConAlleFalse:respond_typeMixed:critical1_new` = b_Intercept + b_respond_typeMixed + b_logConAlleFalse + b_critical2 + `b_logConAlleFalse:critical2` + `b_logConAlleFalse:respond_typeMixed` + `b_respond_typeMixed:critical2` + `b_logConAlleFalse:respond_typeMixed:critical2`,
         `b_logConAlleFalse:respond_typeMixed:critical2_new` = b_Intercept + b_respond_typeMixed + b_logConAlleFalse + b_critical1 + `b_logConAlleFalse:critical1` + `b_logConAlleFalse:respond_typeMixed` + `b_respond_typeMixed:critical1` + `b_logConAlleFalse:respond_typeMixed:critical1`) %>%
  gather_draws(`b_logConAlleBiased:respond_typeMixed_new`,
               `b_logConAlleBiased:respond_typeMixed:critical1_new`, 
               `b_logConAlleBiased:respond_typeMixed:critical2_new`,
               `b_logConAlleUnbiased:respond_typeMixed_new`,
               `b_logConAlleUnbiased:respond_typeMixed:critical1_new`,
               `b_logConAlleUnbiased:respond_typeMixed:critical2_new`,
               `b_logConAlleFalse:respond_typeMixed_new`,
               `b_logConAlleFalse:respond_typeMixed:critical1_new`, 
               `b_logConAlleFalse:respond_typeMixed:critical2_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_logConAlleBiased:respond_typeMixed_new",
               "b_logConAlleBiased:respond_typeMixed:critical1_new", 
               "b_logConAlleBiased:respond_typeMixed:critical2_new",
               "b_logConAlleUnbiased:respond_typeMixed_new",
               "b_logConAlleUnbiased:respond_typeMixed:critical1_new",
               "b_logConAlleUnbiased:respond_typeMixed:critical2_new",
               "b_logConAlleFalse:respond_typeMixed_new",
               "b_logConAlleFalse:respond_typeMixed:critical1_new", 
               "b_logConAlleFalse:respond_typeMixed:critical2_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  hjust = .5,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_blank()) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(3,3,3,3)),
                    breaks = c("b_logConAlleBiased:respond_typeMixed_new",
                               "b_logConAlleUnbiased:respond_typeMixed_new",
                               "b_logConAlleFalse:respond_typeMixed_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False")) +
  ggtitle("\nMixed")

m_crit_alle_prag <- RT_SHAPE_respond_crit %>%
  spread_draws(b_Intercept,
               b_critical1,
               b_critical2,
               b_respond_typePragmatic,
               `b_respond_typePragmatic:critical1`,
               `b_respond_typePragmatic:critical2`,
               b_logConAlleBiased,
               `b_logConAlleBiased:critical1`, 
               `b_logConAlleBiased:critical2`,
               `b_logConAlleBiased:respond_typePragmatic`,
               `b_logConAlleBiased:respond_typePragmatic:critical1`, 
               `b_logConAlleBiased:respond_typePragmatic:critical2`,
               b_logConAlleUnbiased,
               `b_logConAlleUnbiased:critical1`,
               `b_logConAlleUnbiased:critical2`,
               `b_logConAlleUnbiased:respond_typePragmatic`,
               `b_logConAlleUnbiased:respond_typePragmatic:critical1`,
               `b_logConAlleUnbiased:respond_typePragmatic:critical2`,
               b_logConAlleFalse,
               `b_logConAlleFalse:critical1`, 
               `b_logConAlleFalse:critical2`,
               `b_logConAlleFalse:respond_typePragmatic`,
               `b_logConAlleFalse:respond_typePragmatic:critical1`, 
               `b_logConAlleFalse:respond_typePragmatic:critical2`) %>%
  mutate(`b_logConAlleBiased:respond_typePragmatic_new` = b_Intercept + b_logConAlleBiased + b_respond_typePragmatic + `b_logConAlleBiased:respond_typePragmatic`,
         `b_logConAlleBiased:respond_typePragmatic:critical1_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleBiased + b_critical2 + `b_logConAlleBiased:respond_typePragmatic` + `b_logConAlleBiased:critical2` + `b_respond_typePragmatic:critical2` + `b_logConAlleBiased:respond_typePragmatic:critical2`,
         `b_logConAlleBiased:respond_typePragmatic:critical2_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleBiased + b_critical1 + `b_logConAlleBiased:respond_typePragmatic` + `b_logConAlleBiased:critical1` + `b_respond_typePragmatic:critical1` + `b_logConAlleBiased:respond_typePragmatic:critical1`,
         `b_logConAlleUnbiased:respond_typePragmatic_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleUnbiased + `b_logConAlleUnbiased:respond_typePragmatic`,
         `b_logConAlleUnbiased:respond_typePragmatic:critical1_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleUnbiased + b_critical2 + `b_logConAlleUnbiased:critical2` + `b_logConAlleUnbiased:respond_typePragmatic` + `b_respond_typePragmatic:critical2` + `b_logConAlleUnbiased:respond_typePragmatic:critical2`,
         `b_logConAlleUnbiased:respond_typePragmatic:critical2_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleUnbiased + b_critical1 + `b_logConAlleUnbiased:critical1` + `b_logConAlleUnbiased:respond_typePragmatic` + `b_respond_typePragmatic:critical1` + `b_logConAlleUnbiased:respond_typePragmatic:critical1`,
         `b_logConAlleFalse:respond_typePragmatic_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleFalse + `b_logConAlleFalse:respond_typePragmatic`,
         `b_logConAlleFalse:respond_typePragmatic:critical1_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleFalse + b_critical2 + `b_logConAlleFalse:critical2` + `b_logConAlleFalse:respond_typePragmatic` + `b_respond_typePragmatic:critical2` + `b_logConAlleFalse:respond_typePragmatic:critical2`,
         `b_logConAlleFalse:respond_typePragmatic:critical2_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleFalse + b_critical1 + `b_logConAlleFalse:critical1` + `b_logConAlleFalse:respond_typePragmatic` + `b_respond_typePragmatic:critical1` + `b_logConAlleFalse:respond_typePragmatic:critical1`) %>%
  gather_draws(`b_logConAlleBiased:respond_typePragmatic_new`,
               `b_logConAlleBiased:respond_typePragmatic:critical1_new`, 
               `b_logConAlleBiased:respond_typePragmatic:critical2_new`,
               `b_logConAlleUnbiased:respond_typePragmatic_new`,
               `b_logConAlleUnbiased:respond_typePragmatic:critical1_new`,
               `b_logConAlleUnbiased:respond_typePragmatic:critical2_new`,
               `b_logConAlleFalse:respond_typePragmatic_new`,
               `b_logConAlleFalse:respond_typePragmatic:critical1_new`, 
               `b_logConAlleFalse:respond_typePragmatic:critical2_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_logConAlleBiased:respond_typePragmatic_new",
               "b_logConAlleBiased:respond_typePragmatic:critical1_new", 
               "b_logConAlleBiased:respond_typePragmatic:critical2_new",
               "b_logConAlleUnbiased:respond_typePragmatic_new",
               "b_logConAlleUnbiased:respond_typePragmatic:critical1_new",
               "b_logConAlleUnbiased:respond_typePragmatic:critical2_new",
               "b_logConAlleFalse:respond_typePragmatic_new",
               "b_logConAlleFalse:respond_typePragmatic:critical1_new", 
               "b_logConAlleFalse:respond_typePragmatic:critical2_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  scale_y_discrete(labels = c('Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate',
                              'Dreiecke', 
                              'Kreise', 
                              'Quadrate'),
                   position = "right") +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  hjust = 1,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(3,3,3,3)),
                    breaks = c("b_logConAlleBiased:respond_typeMixed_new",
                               "b_logConAlleUnbiased:respond_typeMixed_new",
                               "b_logConAlleFalse:respond_typeMixed_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False")) +
  ggtitle("\nPragmatic")

`est-crit-alle` <-
  ggarrange(m_crit_alle_sem, m_crit_alle_mix, m_crit_alle_prag, ncol = 3)

# ggsave("est-crit-alle.png", height = 6, width = 10, dpi = "retina")

`est-crit-respond` <- 
  ggarrange(`est-crit-einige`, `est-crit-alle`, nrow = 2)

`est-crit-respond`
 
 # ggsave("est-crit-respond.png", height = 12, width = 10, dpi = "retina")

```

#### Model 3 (condition x pragmatic profile x block)

```{r fig.align='center', fig.height=10, fig.width=10, echo=FALSE, warning=FALSE, message=FALSE}

m_trial_einige_sem <- RT_SHAPE_respond_trial %>%
  spread_draws(b_Intercept,
               b_trial_nametask_three_2,
               b_trial_nametask_three_3,
               b_trial_nametask_three_4,
               b_logConEinigeUnbiased,
               `b_logConEinigeUnbiased:trial_nametask_three_2`,
               `b_logConEinigeUnbiased:trial_nametask_three_3`,
               `b_logConEinigeUnbiased:trial_nametask_three_4`,
               b_logConEinigeInfelict,
               `b_logConEinigeInfelict:trial_nametask_three_2`, 
               `b_logConEinigeInfelict:trial_nametask_three_3`,
               `b_logConEinigeInfelict:trial_nametask_three_4`,
               b_logConEinigeFalse,
               `b_logConEinigeFalse:trial_nametask_three_2`, 
               `b_logConEinigeFalse:trial_nametask_three_3`,
               `b_logConEinigeFalse:trial_nametask_three_4`) %>%
  mutate(b_trial_nametask_three_2_new = b_Intercept + b_trial_nametask_three_2,
         b_trial_nametask_three_3_new = b_Intercept + b_trial_nametask_three_3,
         b_trial_nametask_three_4_new = b_Intercept + b_trial_nametask_three_4,
         b_logConEinigeUnbiased_new = b_Intercept + b_logConEinigeUnbiased,
         `b_logConEinigeUnbiased:trial_nametask_three_2_new` = b_Intercept + b_logConEinigeUnbiased + b_trial_nametask_three_2 + `b_logConEinigeUnbiased:trial_nametask_three_2`,
         `b_logConEinigeUnbiased:trial_nametask_three_3_new` = b_Intercept + b_logConEinigeUnbiased + b_trial_nametask_three_3 + `b_logConEinigeUnbiased:trial_nametask_three_3`,
         `b_logConEinigeUnbiased:trial_nametask_three_4_new` = b_Intercept + b_logConEinigeUnbiased + b_trial_nametask_three_4 + `b_logConEinigeUnbiased:trial_nametask_three_4`,
         b_logConEinigeInfelict_new = b_Intercept + b_logConEinigeInfelict,
         `b_logConEinigeInfelict:trial_nametask_three_2_new` = b_Intercept + b_logConEinigeInfelict + b_trial_nametask_three_2 + `b_logConEinigeInfelict:trial_nametask_three_2`,
         `b_logConEinigeInfelict:trial_nametask_three_3_new` = b_Intercept + b_logConEinigeInfelict + b_trial_nametask_three_3 + `b_logConEinigeInfelict:trial_nametask_three_3`,
         `b_logConEinigeInfelict:trial_nametask_three_4_new` = b_Intercept + b_logConEinigeInfelict + b_trial_nametask_three_4 + `b_logConEinigeInfelict:trial_nametask_three_4`,
         b_logConEinigeFalse_new = b_Intercept + b_logConEinigeFalse,
         `b_logConEinigeFalse:trial_nametask_three_2_new` = b_Intercept + b_logConEinigeFalse + b_trial_nametask_three_2 + `b_logConEinigeFalse:trial_nametask_three_2`,
         `b_logConEinigeFalse:trial_nametask_three_3_new` = b_Intercept + b_logConEinigeFalse + b_trial_nametask_three_3 + `b_logConEinigeFalse:trial_nametask_three_3`,
         `b_logConEinigeFalse:trial_nametask_three_4_new` = b_Intercept + b_logConEinigeFalse + b_trial_nametask_three_4 + `b_logConEinigeFalse:trial_nametask_three_4`) %>%
  gather_draws(b_Intercept,
               b_trial_nametask_three_2_new,
               b_trial_nametask_three_3_new,
               b_trial_nametask_three_4_new,
               b_logConEinigeUnbiased_new,
               `b_logConEinigeUnbiased:trial_nametask_three_2_new`,
               `b_logConEinigeUnbiased:trial_nametask_three_3_new`,
               `b_logConEinigeUnbiased:trial_nametask_three_4_new`,
               b_logConEinigeFalse_new,
               `b_logConEinigeFalse:trial_nametask_three_2_new`, 
               `b_logConEinigeFalse:trial_nametask_three_3_new`,
               `b_logConEinigeFalse:trial_nametask_three_4_new`,
                b_logConEinigeInfelict_new,
               `b_logConEinigeInfelict:trial_nametask_three_2_new`, 
               `b_logConEinigeInfelict:trial_nametask_three_3_new`,
               `b_logConEinigeInfelict:trial_nametask_three_4_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_Intercept",
               "b_trial_nametask_three_2_new",
               "b_trial_nametask_three_3_new",
               "b_trial_nametask_three_4_new",
               "b_logConEinigeUnbiased_new",
               "b_logConEinigeUnbiased:trial_nametask_three_2_new",
               "b_logConEinigeUnbiased:trial_nametask_three_3_new",
               "b_logConEinigeUnbiased:trial_nametask_three_4_new",
               "b_logConEinigeFalse_new",
               "b_logConEinigeFalse:trial_nametask_three_2_new", 
               "b_logConEinigeFalse:trial_nametask_three_3_new",
               "b_logConEinigeFalse:trial_nametask_three_4_new",
               "b_logConEinigeInfelict_new",
               "b_logConEinigeInfelict:trial_nametask_three_2_new", 
               "b_logConEinigeInfelict:trial_nametask_three_3_new",
               "b_logConEinigeInfelict:trial_nametask_three_4_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  scale_y_discrete(labels = c('Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84')) +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(4,4,4,4)),
                    breaks = c("b_Intercept",
                               "b_logConEinigeUnbiased_new",
                               "b_logConEinigeFalse_new",
                               "b_logConEinigeInfelict_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False",
                               "Infelicitous")) +
  ggtitle("Einige\nSemantic")

m_trial_einige_mix <- RT_SHAPE_respond_trial %>%
  spread_draws(b_Intercept,
               b_trial_nametask_three_2,
               b_trial_nametask_three_3,
               b_trial_nametask_three_4,
               b_respond_typeMixed,
               `b_respond_typeMixed:trial_nametask_three_2`,
               `b_respond_typeMixed:trial_nametask_three_3`,
               `b_respond_typeMixed:trial_nametask_three_4`,
               b_logConEinigeUnbiased,
               `b_logConEinigeUnbiased:trial_nametask_three_2`,
               `b_logConEinigeUnbiased:trial_nametask_three_3`,
               `b_logConEinigeUnbiased:trial_nametask_three_4`,
               `b_logConEinigeUnbiased:respond_typeMixed`,
               `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_2`,
               `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_3`,
               `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_4`,
               b_logConEinigeInfelict,
               `b_logConEinigeInfelict:trial_nametask_three_2`,
               `b_logConEinigeInfelict:trial_nametask_three_3`,
               `b_logConEinigeInfelict:trial_nametask_three_4`,
               `b_logConEinigeInfelict:respond_typeMixed`,
               `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_2`, 
               `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_3`,
               `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_4`,
               b_logConEinigeFalse,
               `b_logConEinigeFalse:trial_nametask_three_2`,
               `b_logConEinigeFalse:trial_nametask_three_3`,
               `b_logConEinigeFalse:trial_nametask_three_4`,
               `b_logConEinigeFalse:respond_typeMixed`,
               `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_2`, 
               `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_3`,
               `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_4`) %>%
  mutate(b_respond_typeMixed_new = b_Intercept + b_respond_typeMixed,
         `b_respond_typeMixed:trial_nametask_three_2_new` = b_Intercept + b_respond_typeMixed + b_trial_nametask_three_2 + `b_respond_typeMixed:trial_nametask_three_2`,
         `b_respond_typeMixed:trial_nametask_three_3_new` = b_Intercept + b_respond_typeMixed + b_trial_nametask_three_3 + `b_respond_typeMixed:trial_nametask_three_3`,
         `b_respond_typeMixed:trial_nametask_three_4_new` = b_Intercept + b_respond_typeMixed + b_trial_nametask_three_4 + `b_respond_typeMixed:trial_nametask_three_4`,
         `b_logConEinigeUnbiased:respond_typeMixed_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeUnbiased + `b_logConEinigeUnbiased:respond_typeMixed`,
         `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_2_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeUnbiased + b_trial_nametask_three_2 + `b_logConEinigeUnbiased:trial_nametask_three_2` + `b_logConEinigeUnbiased:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_2` + `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_2`,
         `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_3_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeUnbiased + b_trial_nametask_three_3 + `b_logConEinigeUnbiased:trial_nametask_three_3` + `b_logConEinigeUnbiased:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_3` + `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_3`,
         `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_4_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeUnbiased + b_trial_nametask_three_4 + `b_logConEinigeUnbiased:trial_nametask_three_4` + `b_logConEinigeUnbiased:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_4` + `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_4`,
         `b_logConEinigeInfelict:respond_typeMixed_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeInfelict + `b_logConEinigeInfelict:respond_typeMixed`,
         `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_2_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeInfelict + b_trial_nametask_three_2 + `b_logConEinigeInfelict:trial_nametask_three_2` + `b_logConEinigeInfelict:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_2` + `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_2`,
         `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_3_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeInfelict + b_trial_nametask_three_3 + `b_logConEinigeInfelict:trial_nametask_three_3` + `b_logConEinigeInfelict:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_3` + `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_3`,
         `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_4_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeInfelict + b_trial_nametask_three_4 + `b_logConEinigeInfelict:trial_nametask_three_4` + `b_logConEinigeInfelict:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_4` + `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_4`,
         `b_logConEinigeFalse:respond_typeMixed_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeFalse + `b_logConEinigeFalse:respond_typeMixed`,
         `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_2_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeFalse + b_trial_nametask_three_2 + `b_logConEinigeFalse:trial_nametask_three_2` + `b_logConEinigeFalse:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_2` + `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_2`,
         `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_3_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeFalse + b_trial_nametask_three_3 + `b_logConEinigeFalse:trial_nametask_three_3` + `b_logConEinigeFalse:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_3` + `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_3`,
         `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_4_new` = b_Intercept + b_respond_typeMixed + b_logConEinigeFalse + b_trial_nametask_three_4 + `b_logConEinigeFalse:trial_nametask_three_4` + `b_logConEinigeFalse:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_4` + `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_4`) %>%
  gather_draws(b_respond_typeMixed_new,
               `b_respond_typeMixed:trial_nametask_three_2_new`,
               `b_respond_typeMixed:trial_nametask_three_3_new`,
               `b_respond_typeMixed:trial_nametask_three_4_new`,
               `b_logConEinigeUnbiased:respond_typeMixed_new`,
               `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_2_new`,
               `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_3_new`,
               `b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_4_new`,
               `b_logConEinigeFalse:respond_typeMixed_new`,
               `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_2_new`,
               `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_3_new`,
               `b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_4_new`,
               `b_logConEinigeInfelict:respond_typeMixed_new`,
               `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_2_new`, 
               `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_3_new`,
               `b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_4_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_respond_typeMixed_new",
               "b_respond_typeMixed:trial_nametask_three_2_new",
               "b_respond_typeMixed:trial_nametask_three_3_new",
               "b_respond_typeMixed:trial_nametask_three_4_new",
               "b_logConEinigeUnbiased:respond_typeMixed_new",
               "b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_2_new",
               "b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_3_new",
               "b_logConEinigeUnbiased:respond_typeMixed:trial_nametask_three_4_new",
               "b_logConEinigeFalse:respond_typeMixed_new",
               "b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_2_new", 
               "b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_3_new",
               "b_logConEinigeFalse:respond_typeMixed:trial_nametask_three_4_new",
               "b_logConEinigeInfelict:respond_typeMixed_new",
               "b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_2_new", 
               "b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_3_new",
               "b_logConEinigeInfelict:respond_typeMixed:trial_nametask_three_4_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  hjust = .5,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_blank()) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(4,4,4,4)),
                    breaks = c("b_respond_typeMixed_new",
                               "b_logConEinigeUnbiased:respond_typeMixed_new",
                               "b_logConEinigeFalse:respond_typeMixed_new",
                               "b_logConEinigeInfelict:respond_typeMixed_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False",
                               "Infelicitous")) +
  ggtitle("\nMixed")

m_trial_einige_prag <- RT_SHAPE_respond_trial %>%
  spread_draws(b_Intercept,
               b_trial_nametask_three_2,
               b_trial_nametask_three_3,
               b_trial_nametask_three_4,
               b_respond_typePragmatic,
               `b_respond_typePragmatic:trial_nametask_three_2`,
               `b_respond_typePragmatic:trial_nametask_three_3`,
               `b_respond_typePragmatic:trial_nametask_three_4`,
               b_logConEinigeUnbiased,
               `b_logConEinigeUnbiased:trial_nametask_three_2`,
               `b_logConEinigeUnbiased:trial_nametask_three_3`,
               `b_logConEinigeUnbiased:trial_nametask_three_4`,
               `b_logConEinigeUnbiased:respond_typePragmatic`,
               `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_2`,
               `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_3`,
               `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_4`,
               b_logConEinigeInfelict,
               `b_logConEinigeInfelict:trial_nametask_three_2`, 
               `b_logConEinigeInfelict:trial_nametask_three_3`,
               `b_logConEinigeInfelict:trial_nametask_three_4`,
               `b_logConEinigeInfelict:respond_typePragmatic`,
               `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_2`, 
               `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_3`,
               `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_4`,
               b_logConEinigeFalse,
               `b_logConEinigeFalse:trial_nametask_three_2`, 
               `b_logConEinigeFalse:trial_nametask_three_3`,
               `b_logConEinigeFalse:trial_nametask_three_4`,
               `b_logConEinigeFalse:respond_typePragmatic`,
               `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_2`, 
               `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_3`,
               `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_4`) %>%
  mutate(b_respond_typePragmatic_new = b_Intercept + b_respond_typePragmatic,
         `b_respond_typePragmatic:trial_nametask_three_2_new` = b_Intercept + b_respond_typePragmatic + b_trial_nametask_three_2 + `b_respond_typePragmatic:trial_nametask_three_2`,
         `b_respond_typePragmatic:trial_nametask_three_3_new` = b_Intercept + b_respond_typePragmatic + b_trial_nametask_three_3 + `b_respond_typePragmatic:trial_nametask_three_3`,
         `b_respond_typePragmatic:trial_nametask_three_4_new` = b_Intercept + b_respond_typePragmatic + b_trial_nametask_three_4 + `b_respond_typePragmatic:trial_nametask_three_4`,
         `b_logConEinigeUnbiased:respond_typePragmatic_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeUnbiased + `b_logConEinigeUnbiased:respond_typePragmatic`,
         `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_2_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeUnbiased + b_trial_nametask_three_2 + `b_logConEinigeUnbiased:trial_nametask_three_2` + `b_logConEinigeUnbiased:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_2` + `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_2`,
         `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_3_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeUnbiased + b_trial_nametask_three_3 + `b_logConEinigeUnbiased:trial_nametask_three_3` + `b_logConEinigeUnbiased:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_3` + `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_3`,
         `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_4_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeUnbiased + b_trial_nametask_three_4 + `b_logConEinigeUnbiased:trial_nametask_three_4` + `b_logConEinigeUnbiased:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_4` + `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_4`,
         `b_logConEinigeInfelict:respond_typePragmatic_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeInfelict + `b_logConEinigeInfelict:respond_typePragmatic`,
         `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_2_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeInfelict + b_trial_nametask_three_2 + `b_logConEinigeInfelict:trial_nametask_three_2` + `b_logConEinigeInfelict:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_2` + `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_2`,
         `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_3_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeInfelict + b_trial_nametask_three_3 + `b_logConEinigeInfelict:trial_nametask_three_3` + `b_logConEinigeInfelict:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_3` + `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_3`,
         `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_4_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeInfelict + b_trial_nametask_three_4 + `b_logConEinigeInfelict:trial_nametask_three_4` + `b_logConEinigeInfelict:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_4` + `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_4`,
         `b_logConEinigeFalse:respond_typePragmatic_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeFalse + `b_logConEinigeFalse:respond_typePragmatic`,
         `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_2_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeFalse + b_trial_nametask_three_2 + `b_logConEinigeFalse:trial_nametask_three_2` + `b_logConEinigeFalse:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_2` + `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_2`,
         `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_3_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeFalse + b_trial_nametask_three_3 + `b_logConEinigeFalse:trial_nametask_three_3` + `b_logConEinigeFalse:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_3` + `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_3`,
         `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_4_new` = b_Intercept + b_respond_typePragmatic + b_logConEinigeFalse + b_trial_nametask_three_4 + `b_logConEinigeFalse:trial_nametask_three_4` + `b_logConEinigeFalse:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_4` + `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_4`) %>%
  gather_draws(b_respond_typePragmatic_new,
               `b_respond_typePragmatic:trial_nametask_three_2_new`,
               `b_respond_typePragmatic:trial_nametask_three_3_new`,
               `b_respond_typePragmatic:trial_nametask_three_4_new`,
               `b_logConEinigeUnbiased:respond_typePragmatic_new`,
               `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_2_new`,
               `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_3_new`,
               `b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_4_new`,
               `b_logConEinigeFalse:respond_typePragmatic_new`,
               `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_2_new`, 
               `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_3_new`,
               `b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_4_new`,
               `b_logConEinigeInfelict:respond_typePragmatic_new`,
               `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_2_new`, 
               `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_3_new`,
               `b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_4_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_respond_typePragmatic_new",
               "b_respond_typePragmatic:trial_nametask_three_2_new",
               "b_respond_typePragmatic:trial_nametask_three_3_new",
               "b_respond_typePragmatic:trial_nametask_three_4_new",
               "b_logConEinigeUnbiased:respond_typePragmatic_new",
               "b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_2_new",
               "b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_3_new",
               "b_logConEinigeUnbiased:respond_typePragmatic:trial_nametask_three_4_new",
               "b_logConEinigeFalse:respond_typePragmatic_new",
               "b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_2_new",
               "b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_3_new",
               "b_logConEinigeFalse:respond_typePragmatic:trial_nametask_three_4_new",
               "b_logConEinigeInfelict:respond_typePragmatic_new",
               "b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_2_new", 
               "b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_3_new",
               "b_logConEinigeInfelict:respond_typePragmatic:trial_nametask_three_4_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  scale_y_discrete(labels = c('Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84'),
                   position = "right") +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  hjust = 1,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(4,4,4,4)),
                    breaks = c("b_respond_typePragmatic_new",
                               "b_logConEinigeUnbiased:respond_typePragmatic_new",
                               "b_logConEinigeFalse:respond_typePragmatic_new",
                               "b_logConEinigeInfelict:respond_typePragmatic_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False",
                               "Infelicitous")) +
  ggtitle("\nPragmatic")

`est-block-einige` <-
  ggarrange(m_trial_einige_sem, m_trial_einige_mix, m_trial_einige_prag, common.legend = T, legend = "bottom", ncol = 3)

# ggsave("est-block-einige.png", height = 6, width = 10, dpi = "retina")

```


```{r fig.align='center', fig.height=10, fig.width=10, echo=FALSE, warning=FALSE, message=FALSE}

m_trial_alle_sem <- RT_SHAPE_respond_trial %>%
  spread_draws(b_Intercept,
               b_trial_nametask_three_2,
               b_trial_nametask_three_3,
               b_trial_nametask_three_4,
               b_logConAlleBiased,
               `b_logConAlleBiased:trial_nametask_three_2`,
               `b_logConAlleBiased:trial_nametask_three_3`,
               `b_logConAlleBiased:trial_nametask_three_4`,
               b_logConAlleUnbiased,
               `b_logConAlleUnbiased:trial_nametask_three_2`,
               `b_logConAlleUnbiased:trial_nametask_three_3`,
               `b_logConAlleUnbiased:trial_nametask_three_4`,
               b_logConAlleFalse,
               `b_logConAlleFalse:trial_nametask_three_2`, 
               `b_logConAlleFalse:trial_nametask_three_3`,
               `b_logConAlleFalse:trial_nametask_three_4`) %>%
  mutate(b_logConAlleBiased_new = b_Intercept + b_logConAlleBiased,
         `b_logConAlleBiased:trial_nametask_three_2_new` = b_Intercept + b_logConAlleBiased + b_trial_nametask_three_2 + `b_logConAlleBiased:trial_nametask_three_2`,
         `b_logConAlleBiased:trial_nametask_three_3_new` = b_Intercept + b_logConAlleBiased + b_trial_nametask_three_3 + `b_logConAlleBiased:trial_nametask_three_3`,
         `b_logConAlleBiased:trial_nametask_three_4_new` = b_Intercept + b_logConAlleBiased + b_trial_nametask_three_4 + `b_logConAlleBiased:trial_nametask_three_4`,
         b_logConAlleUnbiased_new = b_Intercept + b_logConAlleUnbiased,
         `b_logConAlleUnbiased:trial_nametask_three_2_new` = b_Intercept + b_logConAlleUnbiased + b_trial_nametask_three_2 + `b_logConAlleUnbiased:trial_nametask_three_2`,
         `b_logConAlleUnbiased:trial_nametask_three_3_new` = b_Intercept + b_logConAlleUnbiased + b_trial_nametask_three_3 + `b_logConAlleUnbiased:trial_nametask_three_3`,
         `b_logConAlleUnbiased:trial_nametask_three_4_new` = b_Intercept + b_logConAlleUnbiased + b_trial_nametask_three_4 + `b_logConAlleUnbiased:trial_nametask_three_4`,
         b_logConAlleFalse_new = b_Intercept + b_logConAlleFalse,
         `b_logConAlleFalse:trial_nametask_three_2_new` = b_Intercept + b_logConAlleFalse + b_trial_nametask_three_2 + `b_logConAlleFalse:trial_nametask_three_2`,
         `b_logConAlleFalse:trial_nametask_three_3_new` = b_Intercept + b_logConAlleFalse + b_trial_nametask_three_3 + `b_logConAlleFalse:trial_nametask_three_3`,
         `b_logConAlleFalse:trial_nametask_three_4_new` = b_Intercept + b_logConAlleFalse + b_trial_nametask_three_4 + `b_logConAlleFalse:trial_nametask_three_4`) %>%
  gather_draws(b_logConAlleBiased_new,
               `b_logConAlleBiased:trial_nametask_three_2_new`,
               `b_logConAlleBiased:trial_nametask_three_3_new`,
               `b_logConAlleBiased:trial_nametask_three_4_new`,
               b_logConAlleUnbiased_new,
               `b_logConAlleUnbiased:trial_nametask_three_2_new`,
               `b_logConAlleUnbiased:trial_nametask_three_3_new`,
               `b_logConAlleUnbiased:trial_nametask_three_4_new`,
               b_logConAlleFalse_new,
               `b_logConAlleFalse:trial_nametask_three_2_new`, 
               `b_logConAlleFalse:trial_nametask_three_3_new`,
               `b_logConAlleFalse:trial_nametask_three_4_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_logConAlleBiased_new",
               "b_logConAlleBiased:trial_nametask_three_2_new",
               "b_logConAlleBiased:trial_nametask_three_3_new",
               "b_logConAlleBiased:trial_nametask_three_4_new",
               "b_logConAlleUnbiased_new",
               "b_logConAlleUnbiased:trial_nametask_three_2_new",
               "b_logConAlleUnbiased:trial_nametask_three_3_new",
               "b_logConAlleUnbiased:trial_nametask_three_4_new",
               "b_logConAlleFalse_new",
               "b_logConAlleFalse:trial_nametask_three_2_new", 
               "b_logConAlleFalse:trial_nametask_three_3_new",
               "b_logConAlleFalse:trial_nametask_three_4_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  scale_y_discrete(labels = c('Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84')) +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(4,4,4,4)),
                    breaks = c("b_logConAlleBiased_new",
                               "b_logConAlleUnbiased_new",
                               "b_logConAlleFalse_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False")) +
  ggtitle("Alle\nSemantic")

m_trial_alle_mix <- RT_SHAPE_respond_trial %>%
  spread_draws(b_Intercept,
               b_trial_nametask_three_2,
               b_trial_nametask_three_3,
               b_trial_nametask_three_4,
               b_respond_typeMixed,
               `b_respond_typeMixed:trial_nametask_three_2`,
               `b_respond_typeMixed:trial_nametask_three_3`,
               `b_respond_typeMixed:trial_nametask_three_4`,
               b_logConAlleBiased,
               `b_logConAlleBiased:trial_nametask_three_2`, 
               `b_logConAlleBiased:trial_nametask_three_3`,
               `b_logConAlleBiased:trial_nametask_three_4`,
               `b_logConAlleBiased:respond_typeMixed`,
               `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_2`, 
               `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_3`,
               `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_4`,
               b_logConAlleUnbiased,
               `b_logConAlleUnbiased:trial_nametask_three_2`,
               `b_logConAlleUnbiased:trial_nametask_three_3`,
               `b_logConAlleUnbiased:trial_nametask_three_4`,
               `b_logConAlleUnbiased:respond_typeMixed`,
               `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_2`,
               `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_3`,
               `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_4`,
               b_logConAlleFalse,
               `b_logConAlleFalse:trial_nametask_three_2`, 
               `b_logConAlleFalse:trial_nametask_three_3`,
               `b_logConAlleFalse:trial_nametask_three_4`,
               `b_logConAlleFalse:respond_typeMixed`,
               `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_2`, 
               `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_3`,
               `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_4`) %>%
  mutate(`b_logConAlleBiased:respond_typeMixed_new` = b_Intercept + b_logConAlleBiased + b_respond_typeMixed + `b_logConAlleBiased:respond_typeMixed`,
         `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_2_new` = b_Intercept + b_respond_typeMixed + b_logConAlleBiased + b_trial_nametask_three_2 + `b_logConAlleBiased:respond_typeMixed` + `b_logConAlleBiased:trial_nametask_three_2` + `b_respond_typeMixed:trial_nametask_three_2` + `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_2`,
         `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_3_new` = b_Intercept + b_respond_typeMixed + b_logConAlleBiased + b_trial_nametask_three_3 + `b_logConAlleBiased:respond_typeMixed` + `b_logConAlleBiased:trial_nametask_three_3` + `b_respond_typeMixed:trial_nametask_three_3` + `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_3`,
         `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_4_new` = b_Intercept + b_respond_typeMixed + b_logConAlleBiased + b_trial_nametask_three_4 + `b_logConAlleBiased:respond_typeMixed` + `b_logConAlleBiased:trial_nametask_three_4` + `b_respond_typeMixed:trial_nametask_three_4` + `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_4`,
         `b_logConAlleUnbiased:respond_typeMixed_new` = b_Intercept + b_respond_typeMixed + b_logConAlleUnbiased + `b_logConAlleUnbiased:respond_typeMixed`,
         `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_2_new` = b_Intercept + b_respond_typeMixed + b_logConAlleUnbiased + b_trial_nametask_three_2 + `b_logConAlleUnbiased:trial_nametask_three_2` + `b_logConAlleUnbiased:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_2` + `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_2`,
         `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_3_new` = b_Intercept + b_respond_typeMixed + b_logConAlleUnbiased + b_trial_nametask_three_3 + `b_logConAlleUnbiased:trial_nametask_three_3` + `b_logConAlleUnbiased:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_3` + `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_3`,
         `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_4_new` = b_Intercept + b_respond_typeMixed + b_logConAlleUnbiased + b_trial_nametask_three_4 + `b_logConAlleUnbiased:trial_nametask_three_4` + `b_logConAlleUnbiased:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_4` + `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_4`,
         `b_logConAlleFalse:respond_typeMixed_new` = b_Intercept + b_respond_typeMixed + b_logConAlleFalse + `b_logConAlleFalse:respond_typeMixed`,
         `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_2_new` = b_Intercept + b_respond_typeMixed + b_logConAlleFalse + b_trial_nametask_three_2 + `b_logConAlleFalse:trial_nametask_three_2` + `b_logConAlleFalse:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_2` + `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_2`,
         `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_3_new` = b_Intercept + b_respond_typeMixed + b_logConAlleFalse + b_trial_nametask_three_3 + `b_logConAlleFalse:trial_nametask_three_3` + `b_logConAlleFalse:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_3` + `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_3`,
         `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_4_new` = b_Intercept + b_respond_typeMixed + b_logConAlleFalse + b_trial_nametask_three_4 + `b_logConAlleFalse:trial_nametask_three_4` + `b_logConAlleFalse:respond_typeMixed` + `b_respond_typeMixed:trial_nametask_three_4` + `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_4`) %>%
  gather_draws(`b_logConAlleBiased:respond_typeMixed_new`,
               `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_2_new`, 
               `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_3_new`,
               `b_logConAlleBiased:respond_typeMixed:trial_nametask_three_4_new`,
               `b_logConAlleUnbiased:respond_typeMixed_new`,
               `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_2_new`,
               `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_3_new`,
               `b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_4_new`,
               `b_logConAlleFalse:respond_typeMixed_new`,
               `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_2_new`, 
               `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_3_new`,
               `b_logConAlleFalse:respond_typeMixed:trial_nametask_three_4_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_logConAlleBiased:respond_typeMixed_new",
               "b_logConAlleBiased:respond_typeMixed:trial_nametask_three_2_new", 
               "b_logConAlleBiased:respond_typeMixed:trial_nametask_three_3_new",
               "b_logConAlleBiased:respond_typeMixed:trial_nametask_three_4_new",
               "b_logConAlleUnbiased:respond_typeMixed_new",
               "b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_2_new",
               "b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_3_new",
               "b_logConAlleUnbiased:respond_typeMixed:trial_nametask_three_4_new",
               "b_logConAlleFalse:respond_typeMixed_new",
               "b_logConAlleFalse:respond_typeMixed:trial_nametask_three_2_new", 
               "b_logConAlleFalse:respond_typeMixed:trial_nametask_three_3_new",
               "b_logConAlleFalse:respond_typeMixed:trial_nametask_three_4_new"))) %>%
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  hjust = .5,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_blank()) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(3,3,3,3)),
                    breaks = c("b_logConAlleBiased:respond_typeMixed_new",
                               "b_logConAlleUnbiased:respond_typeMixed_new",
                               "b_logConAlleFalse:respond_typeMixed_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False")) +
  ggtitle("\nMixed")

m_trial_alle_prag <- RT_SHAPE_respond_trial %>%
  spread_draws(b_Intercept,
               b_trial_nametask_three_2,
               b_trial_nametask_three_3,
               b_trial_nametask_three_4,
               b_respond_typePragmatic,
               `b_respond_typePragmatic:trial_nametask_three_2`,
               `b_respond_typePragmatic:trial_nametask_three_3`,
               `b_respond_typePragmatic:trial_nametask_three_4`,
               b_logConAlleBiased,
               `b_logConAlleBiased:trial_nametask_three_2`, 
               `b_logConAlleBiased:trial_nametask_three_3`,
               `b_logConAlleBiased:trial_nametask_three_4`,
               `b_logConAlleBiased:respond_typePragmatic`,
               `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_2`, 
               `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_3`,
               `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_4`,
               b_logConAlleUnbiased,
               `b_logConAlleUnbiased:trial_nametask_three_2`,
               `b_logConAlleUnbiased:trial_nametask_three_3`,
               `b_logConAlleUnbiased:trial_nametask_three_4`,
               `b_logConAlleUnbiased:respond_typePragmatic`,
               `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_2`,
               `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_3`,
               `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_4`,
               b_logConAlleFalse,
               `b_logConAlleFalse:trial_nametask_three_2`, 
               `b_logConAlleFalse:trial_nametask_three_3`,
               `b_logConAlleFalse:trial_nametask_three_4`,
               `b_logConAlleFalse:respond_typePragmatic`,
               `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_2`, 
               `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_3`,
               `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_4`) %>%
  mutate(`b_logConAlleBiased:respond_typePragmatic_new` = b_Intercept + b_logConAlleBiased + b_respond_typePragmatic + `b_logConAlleBiased:respond_typePragmatic`,
         `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_2_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleBiased + b_trial_nametask_three_2 + `b_logConAlleBiased:respond_typePragmatic` + `b_logConAlleBiased:trial_nametask_three_2` + `b_respond_typePragmatic:trial_nametask_three_2` + `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_2`,
         `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_3_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleBiased + b_trial_nametask_three_3 + `b_logConAlleBiased:respond_typePragmatic` + `b_logConAlleBiased:trial_nametask_three_3` + `b_respond_typePragmatic:trial_nametask_three_3` + `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_3`,
         `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_4_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleBiased + b_trial_nametask_three_4 + `b_logConAlleBiased:respond_typePragmatic` + `b_logConAlleBiased:trial_nametask_three_4` + `b_respond_typePragmatic:trial_nametask_three_4` + `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_4`,
         `b_logConAlleUnbiased:respond_typePragmatic_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleUnbiased + `b_logConAlleUnbiased:respond_typePragmatic`,
         `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_2_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleUnbiased + b_trial_nametask_three_2 + `b_logConAlleUnbiased:trial_nametask_three_2` + `b_logConAlleUnbiased:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_2` + `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_2`,
         `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_3_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleUnbiased + b_trial_nametask_three_3 + `b_logConAlleUnbiased:trial_nametask_three_3` + `b_logConAlleUnbiased:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_3` + `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_3`,
         `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_4_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleUnbiased + b_trial_nametask_three_4 + `b_logConAlleUnbiased:trial_nametask_three_4` + `b_logConAlleUnbiased:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_4` + `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_4`,
         `b_logConAlleFalse:respond_typePragmatic_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleFalse + `b_logConAlleFalse:respond_typePragmatic`,
         `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_2_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleFalse + b_trial_nametask_three_2 + `b_logConAlleFalse:trial_nametask_three_2` + `b_logConAlleFalse:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_2` + `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_2`,
         `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_3_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleFalse + b_trial_nametask_three_3 + `b_logConAlleFalse:trial_nametask_three_3` + `b_logConAlleFalse:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_3` + `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_3`,
         `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_4_new` = b_Intercept + b_respond_typePragmatic + b_logConAlleFalse + b_trial_nametask_three_4 + `b_logConAlleFalse:trial_nametask_three_4` + `b_logConAlleFalse:respond_typePragmatic` + `b_respond_typePragmatic:trial_nametask_three_4` + `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_4`) %>%
  gather_draws(`b_logConAlleBiased:respond_typePragmatic_new`,
               `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_2_new`, 
               `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_3_new`,
               `b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_4_new`,
               `b_logConAlleUnbiased:respond_typePragmatic_new`,
               `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_2_new`,
               `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_3_new`,
               `b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_4_new`,
               `b_logConAlleFalse:respond_typePragmatic_new`,
               `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_2_new`, 
               `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_3_new`,
               `b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_4_new`) %>%
  mutate(.variable = factor(x = .variable, 
         levels = c("b_logConAlleBiased:respond_typePragmatic_new",
               "b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_2_new", 
               "b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_3_new",
               "b_logConAlleBiased:respond_typePragmatic:trial_nametask_three_4_new",
               "b_logConAlleUnbiased:respond_typePragmatic_new",
               "b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_2_new",
               "b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_3_new",
               "b_logConAlleUnbiased:respond_typePragmatic:trial_nametask_three_4_new",
               "b_logConAlleFalse:respond_typePragmatic_new",
               "b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_2_new", 
               "b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_3_new",
               "b_logConAlleFalse:respond_typePragmatic:trial_nametask_three_4_new"))) %>% 
  ggplot(aes(y = .variable, x = .value, fill = .variable)) +
  stat_halfeye(.width = c(.90, .95), alpha = .6) +
  geom_pointinterval(alpha = .4) +
  scale_y_discrete(labels = c('Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84',
                              'Trial 1-21',
                              'Trial 22-42',
                              'Trial 43-63',
                              'Trial 64-84'),
                   position = "right") +
  theme_minimal() +
  theme(axis.title.x = element_blank(), 
        axis.title.y = element_blank(),
        legend.title = element_blank(),
        legend.text = element_text(size = 12),
        plot.title = element_text(face = "bold",
                                  size = 14,
                                  hjust = 1,
                                  margin = margin(t = 0, r = 0, b = 15, l = 0)),
        legend.position = "none",
        panel.grid.major.y = element_blank(), 
        axis.text.x = element_text(face = "bold", 
                                   margin = margin(t = 10, r = 0, b = 15, l = 0)),
        axis.text.y = element_text(face = "bold", 
                                   margin = margin(t = 0, r = 10, b = 0, l = 0))) +
  scale_fill_manual(values = rep(c("#F7882F", "#F7C331", "#B2182B", "#94618E"), 
                                 c(3,3,3,3)),
                    breaks = c("b_logConAlleBiased:respond_typeMixed_new",
                               "b_logConAlleUnbiased:respond_typeMixed_new",
                               "b_logConAlleFalse:respond_typeMixed_new"),
                    labels = c("Biased",
                               "Unbiased",
                               "False")) +
  ggtitle("\nPragmatic")

`est-block-alle` <-
  ggarrange(m_trial_alle_sem, m_trial_alle_mix, m_trial_alle_prag, ncol = 3)

# ggsave("est-block-alle.png", height = 6, width = 10, dpi = "retina")
  
`est-block-respond` <- 
  ggarrange(`est-block-einige`, `est-block-alle`, nrow = 2)

`est-block-respond`
 
 # ggsave("est-block-respond.png", height = 12, width = 10, dpi = "retina")

```
